isorropia: reorganize
