+++ src/Aerosol.jl
@@ -1,5 +1,9 @@
 module Aerosol
 
-# Write your package code here.
+using EarthSciMLBase
+using ModelingToolkit, Catalyst, Unitful
+using IfElse
+
+include("isorropia/isorropia.jl")
 
 end
