--- src/ISORROPIA_II_notebook.jl
@@ -1,2001 +1,0 @@
-#Citation (original paper) - Fountoukis, C., Nenes, A., 2007. ISORROPIA II: a computationally efficient thermodynamic equilibrium model for K+-Ca2+-Mg2+-NH4+-Na+-SO42--NO3--Cl--H2O aerosols. Atmospheric Chemistry and Physics 7, 4639–4659. https://doi.org/10.5194/acp-7-4639-2007
-
-#The present code follows the algorithm devised by Fountoukis and Nenes, 2007 in julia language using various packages such as Catalyst.jl and ModellingToolkit.jl. The model utilizes input values for Temperature (T), Relative Humidity and aqueous concentrations (ug/m3) of Na, H2SO4, NH3, HNO3, HCL, Ca2+, K+ AND Mg2+ and predicts the concentrations of various other chemical species (38 species in total).
-
-### A Pluto.jl notebook ###
-# v0.17.3
-
-using Markdown
-using InteractiveUtils
-
-# ╔═╡ 9014f3c0-5dec-11ec-1d46-efedaead02f9
-using Catalyst, OrdinaryDiffEq, DiffEqBase,Plots,StochasticDiffEq, DifferentialEquations, ModelingToolkit, NonlinearSolve
-
-# ╔═╡ e6cfd4c7-fe1f-4991-b5f7-3eaf157b6df9
-@parameters p1 p2 p3 p4 p5 p6 p7 p8 p9 p10 p11 p12 p13 p14 p15 p16 p17 p18 p19 p20 p21 p22 p23 p24 p25 p26 p27 p28 p29 p30 p31 p32 p33 p34 p35 p36 p37 p38 p39 p40 p41 p42 p43 p44 p45 p46 p47 p48 p49 p50 p51 p52 p53 p54 t
-
-# ╔═╡ 1bbc8b1b-6ecd-453a-a508-b561ca347ec7
-@variables CANO32s(t) CA2aq(t) NO3aq(t) CACL2s(t) CLaq(t) CASO4s(t) SO42aq(t) H2Og(t) K2SO4s(t) Kaq(t) KHSO4s(t) HSO4aq(t) KNO3s(t) KCLs(t) MGSO4s(t) MG2aq(t) MGNO32s(t) MGCL2s(t) Haq(t) NH3g(t) NH3aq(t) NH4aq(t) OHaq(t) HNO3g(t) HNO3aq(t) HCLg(t) HCLaq(t) NA2SO4s(t) NAaq(t) NH42SO4s(t) NH4CLs(t) NANO3s(t) NO3aq(t) NACLs(t) NAHSO4s(t) NH4NO3s(t) NH4HSO4s(t) NH43HSO42s(t) H2Oaq(t)	
-
-# ╔═╡ 70c668e3-3b70-4f22-a9eb-415203032080
-rxs = [Reaction(p1, [CANO32s], [CA2aq,NO3aq], [1], [1,2]) 
-		#Ca(NO3)2 --> Ca2+ + 2NO3-
-			Reaction(p2, [CA2aq,NO3aq], [CANO32s], [1,2], [1]) 
-				#Ca2+ + 2NO3- --> Ca(NO3)2
-			Reaction(p3, [CACL2s], [CA2aq,CLaq], [1], [1,2]) 
-				#CaCl2 --> Ca2+ + 2Cl-
-			Reaction(p4, [CA2aq,CLaq], [CACL2s], [1,2], [1]) 
-				#Ca2+ + 2Cl- --> CaCl2
-			Reaction(p5, [CASO4s], [CA2aq,SO42aq,H2Og], [1], [1,1,2]) 
-				#CaSO4.2H2O --> Ca2+ + SO42- + 2H2O
-			Reaction(p6, [CA2aq,SO42aq,H2Og], [CASO4s], [1,1,2], [1]) 
-				#Ca2+ + SO42- + 2H2O --> CaSO4.2H2O
-			Reaction(p7, [K2SO4s], [Kaq,SO42aq], [1], [1,2]) 
-				#K2SO4 --> 2K+ + SO42-
-			Reaction(p8, [Kaq,SO42aq], [K2SO4s], [1,2], [1]) 
-				#2K+ + SO42- --> K2SO4
-			Reaction(p9, [KHSO4s], [Kaq,HSO4aq], [1], [1,1]) 
-				#KHSO4 --> K+ + HSO4- 
-			Reaction(p10, [Kaq,HSO4aq], [KHSO4s], [1,1], [1]) 
-				#K+ + HSO4- --> KHSO4 
-			Reaction(p11, [KNO3s], [Kaq,NO3aq], [1], [1,1]) 
-				#KNO3 --> K+ + NO3-
-			Reaction(p12, [Kaq,NO3aq], [KNO3s], [1,1], [1]) 
-				#K+ + NO3 --> KNO3 
-			Reaction(p13, [KCLs], [Kaq,CLaq], [1], [1,1]) 
-				#KCl --> K+ + Cl-
-			Reaction(p14, [Kaq,CLaq], [KCLs], [1,1], [1]) 
-				#K+ + Cl- --> KCl 
-			Reaction(p15, [MGSO4s], [MG2aq,SO42aq], [1], [1,1]) 
-				#MgSO4 --> Mg2+ + SO42-
-			Reaction(p16, [MG2aq,SO42aq], [MGSO4s], [1,1], [1]) 
-				#Mg2+ + SO42- --> MgSO4
-			Reaction(p17, [MGNO32s], [MG2aq,NO3aq], [1], [1,2]) 
-				#Mg(NO3)2 --> Mg2+ + 2NO3-
-			Reaction(p18, [MG2aq,NO3aq], [MGNO32s], [1,2], [1]) 
-				#Mg2+ + 2NO3- --> Mg(NO3)2 
-			Reaction(p19, [MGCL2s], [MG2aq,CLaq], [1], [1,2]) 
-				#MgCL2 --> Mg2+ + 2Cl-
-			Reaction(p20, [MG2aq,CLaq], [MGCL2s], [1,2], [1]) 
-				#Mg2+ + 2Cl- --> MgCL2 
-			Reaction(p21, [HSO4aq], [Haq,SO42aq], [1], [1,1]) 
-				#HSO4- --> H+ + SO42-
-			Reaction(p22, [Haq,SO42aq], [HSO4aq], [1,1], [1]) 
-				#H+ + SO42- --> HSO4- 
-			Reaction(p23, [NH3g], [NH3aq], [1], [1]) #NH3 --> NH3
-			Reaction(p24, [NH3aq], [NH3g], [1], [1]) #NH3 --> NH3
-			Reaction(p25, [NH3aq,H2Oaq], [NH4aq,OHaq], [1,1], [1,1]) 
-				#NH3 + H2O --> NH4+ + OH-
-			Reaction(p26, [NH4aq,OHaq], [NH3aq,H2Oaq], [1,1], [1,1]) 
-				#NH4+ + OH- --> NH3 + H2O 
-			Reaction(p27, [HNO3g], [Haq,NO3aq], [1], [1,1]) 
-				#HNO3 --> H+ + NO3-
-			Reaction(p28, [Haq,NO3aq], [HNO3g], [1,1], [1]) 
-				#H+ + NO3- --> HNO3 
-			Reaction(p29, [HNO3g], [HNO3aq], [1], [1]) 
-				#HNO3 --> HNO3
-			Reaction(p30, [HNO3aq], [HNO3g], [1], [1]) 
-				#HNO3 --> HNO3
-			Reaction(p31, [HCLg], [Haq,CLaq], [1], [1,1]) 
-				#HCl --> H+ + Cl-
-			Reaction(p32, [Haq,CLaq], [HCLg], [1,1], [1]) 
-				#H+ + Cl- --> HCl 
-			Reaction(p33, [HCLaq], [HCLg], [1], [1]) 
-				#HCl --> HCl
-			Reaction(p34, [HCLg], [HCLaq], [1], [1]) 
-				#HCl --> HCl		
-			Reaction(p35, [H2Oaq], [Haq,OHaq], [1], [1,1]) 
-				#H20 --> H+ + OH-
-			#Reaction(p36, [Haq,OHaq], [H2Oaq], [1,1], [1]) 
-				#H+ + OH- --> H20 
-			Reaction(p37, [NA2SO4s], [NAaq,SO42aq], [1], [2,1]) 
-				#Na2SO4 --> 2Na+ + SO42-
-			Reaction(p38, [NAaq,SO42aq], [NA2SO4s], [2,1], [1]) 
-				#2Na+ + SO42- --> Na2SO4
-			Reaction(p39, [NH42SO4s], [NH4aq,SO42aq], [1], [2,1]) 
-				#(NH4)2SO4 --> 2NH4+ + SO42-
-			Reaction(p40, [NH4aq,SO42aq], [NH42SO4s], [2,1], [1]) 
-				#2NH4+ + SO42- --> (NH4)2SO4 
-			Reaction(p41, [NH4CLs], [NH3g,SO42aq], [1], [1,1]) 
-				#NH4Cl --> NH3 + HCl
-			#Reaction(p42, [NH3g,SO42aq], [NH4CLs], [1,1], [1]) 
-				#NH3 + HCl --> NH4Cl 
-			Reaction(p43, [NANO3s], [NAaq,NO3aq], [1], [1,1]) 
-				#NaNO3 --> Na+ + NO3-
-			Reaction(p44, [NAaq,NO3aq], [NANO3s], [1,1], [1]) 
-				#Na+ + NO3- --> NaNO3 
-			Reaction(p45, [NACLs], [NAaq,CLaq], [1], [1,1]) 
-				#NaCl --> Na+ + Cl-
-			Reaction(p46, [NAaq,CLaq], [NACLs], [1,1], [1]) 
-				#Na+ + Cl- --> NaCl 
-			Reaction(p47, [NAHSO4s], [NAaq,HSO4aq], [1], [1,1]) 
-				#NaHSO4 --> Na+ + HSO4-
-			Reaction(p48, [NAaq,HSO4aq], [NAHSO4s], [1,1], [1]) 
-				#Na+ + HSO4- --> NaHSO4
-			Reaction(p49, [NH4NO3s], [NH3g,HNO3g], [1], [1,1]) 
-				#NH4NO3 --> NH3 + HNO3
-			#Reaction(p50, [NH3g,HNO3g], [NH4NO3s], [1,1], [1]) 
-				#NH3 + HNO3 --> NH4NO3 
-			Reaction(p51, [NH4HSO4s], [NH4aq,HSO4aq], [1], [1,1]) 
-				#NH4HSO4 --> NH4+ + HSO4-
-			Reaction(p52, [NH4aq,HSO4aq], [NH4HSO4s], [1,1], [1]) 
-				#NH4+ + HSO4- --> NH4HSO4 
-			Reaction(p53, [NH43HSO42s], [NH4aq,HSO4aq,SO42aq], [1], [3,1,1]) 
-				#(NH4)3H(SO4)2 --> 3NH4+ + HSO4- + SO42-
-			Reaction(p54, [NH4aq,HSO4aq,SO42aq], [NH43HSO42s], [3,1,1], [1]) 						#3NH4+ + HSO4- + SO42- --> (NH4)3H(SO4)2 
-
-					
-]
-
-
-# ╔═╡ 2e687027-42ef-4370-b64a-dd083f37c7a5
-@named ISORROPIA = ReactionSystem(rxs, t) #creating a reaction system
-
-# ╔═╡ e1fafac6-3851-41ba-a4d6-2b78d7d40389
-u₀map = [CANO32s => 0.0,  CA2aq => 0.4, NO3aq => 0.0,  CACL2s => 0.0, CLaq => 0.0, CASO4s => 0.0, SO42aq => 0.0, H2Og => 0.0, K2SO4s => 0.0, Kaq => 0.33, KHSO4s => 0.0, HSO4aq => 10.0, KNO3s => 0.0, KCLs => 0.0, MGSO4s => 0.0, MG2aq => 0.0, MGNO32s => 0.0, MGCL2s => 0.0, Haq => 10.0, NH3g => 0.0, NH3aq => 3.4, NH4aq => 0.0, OHaq => 0.0, HNO3g => 0.0, HNO3aq => 2.0, HCLg => 0.0, HCLaq => 0.0, NA2SO4s => 0.0, NAaq => 0.0, NH42SO4s => 0.0, NH4CLs => 0.0, NANO3s => 0.0, NO3aq => 0.0, NACLs => 0.0, NAHSO4s => 0.0, NH4NO3s => 0.0, NH4HSO4s => 0.0, NH43HSO42s => 0.0, H2Oaq => 0.0]
-
-	#Input conditions for model
-	
-
-# ╔═╡ d65aa2b0-166f-467e-b25f-64326be2f22d
-tspan = (0.0, 10.0)
-
-# ╔═╡ 4a708fad-4be6-4e80-9aad-338b2bd746ad
-T_ = 298.15
-
-# ╔═╡ 9e9b70ee-3244-4688-af1c-0e9d96391c95
-RH = 0.700
-
-# ╔═╡ c188c227-32da-457a-bb29-2a7a95d31dce
-	parammap = [p1 => ifelse(RH > 0.4906*exp(-509.4*((1/298.15)-(1/T_))), 6.067*10^(5)*exp((-11.299*((298.15/T_)-1))-(1+log((2.718),(298.15/T_))-(298.15/T_))), 0), #Ca(NO3)2
-				p2 => 1.648*10^(-6)*exp((-11.299*((298.15/T_)-1))+(1+log((2.718),(298.15/T_))-(298.15/T_))),
-				p3 => ifelse(RH > 0.2830*exp(-551.1*((1/298.15)-(1/T_))), 7.974*10^(11)*exp((-14.087*((298.15/T_)-1))-(1+log((2.718),(298.15/T_))-(298.15/T_))), 0), #CaCl2
-				p4 => 1.254^(-12)*exp((-14.087*((298.15/T_)-1))+(1+log((2.718),(298.15/T_))-(298.15/T_))),
-				p5 => ifelse(RH > 0.9700*exp(-((1/298.15)-(1/T_))), 4.319*10^(-5)*exp((((298.15/T_)-1))-(1+log((2.718),(298.15/T_))-(298.15/T_))), 0), #CaSO4
-				p6 => 2.315*10^(4)*exp((((298.15/T_)-1))+(1+log((2.718),(298.15/T_))-(298.15/T_))),
-				p7 => ifelse(RH > 0.9751*exp(-35.6*((1/298.15)-(1/T_))), 1.569*10^(-2)*exp((-9.589*((298.15/T_)-1))-(45.807(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #K2SO4
-				p8 => 6.373*10^(-1)*exp((-9.589*((298.15/T_)-1))+(45.807(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p9 => ifelse(RH > 0.8600*exp(-((1/298.15)-(1/T_))), 24.016*exp((-8.423*((298.15/T_)-1))-(17.964(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #KHSO4
-				p10 => 4.164*10^(-2)*exp((-8.423*((298.15/T_)-1))+(17.964(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p11 => ifelse(RH > 0.9248*exp(-((1/298.15)-(1/T_))), 0.872*exp((14.075*((298.15/T_)-1))-(19.388(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #KNO3
-				p12 => 1.147*exp((14.075*((298.15/T_)-1))+(19.388(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p13 => ifelse(RH > 0.8426*exp(-158.9*((1/298.15)-(1/T_))), 8.680*exp((-6.167*((298.15/T_)-1))-(19.953(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #KCl
-				p14 => 1.152*10^(-1)*exp((-6.167*((298.15/T_)-1))+(19.953(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p15 => ifelse(RH > 0.8613*exp(714.5*((1/298.15)-(1/T_))), 1.079*10^(5)*exp((36.798*((298.15/T_)-1))-((1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #MgSO4
-				p16 => 9.268*10^(-6)*exp((36.798*((298.15/T_)-1))+((1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p17 => ifelse(RH > 0.5400*exp(-230.2*((1/298.15)-(1/T_))), 2.507*10^(15)*exp((-8.754*((298.15/T_)-1))-((1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #Mg(NO3)2
-				p18 => 3.989*10^(-16)*exp((-8.754*((298.15/T_)-1))+((1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p19 => ifelse(RH > 0.3284*exp(-42.23*((1/298.15)-(1/T_))), 9.557*10^(21)*exp((-1.347*((298.15/T_)-1))-((1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #MgCl2
-				p20 => 1.046*10^(-22)*exp((-1.347*((298.15/T_)-1))+((1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p21 => 1.015*10^(-2)*exp((8.85*((298.15/T_)-1))-(25.14(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p22 => 9.852*10^(1)*exp((8.85*((298.15/T_)-1))+(25.14(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p23 => 5.764*10^(1)*exp((13.79*((298.15/T_)-1))-(-5.39(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p24 => 1.735*10^(-2)*exp((13.79*((298.15/T_)-1))+(-5.39(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p25 => 1.805*10^(-5)*exp((-1.50*((298.15/T_)-1))-(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p26 => 5.540*10^(4)*exp((-1.50*((298.15/T_)-1))+(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p27 => 2.511*10^(6)*exp((29.17*((298.15/T_)-1))-(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p28 => 3.982*10^(-7)*exp((29.17*((298.15/T_)-1))+(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p29 => 2.1*10^(5)*exp((29.17*((298.15/T_)-1))-(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p30 => 4.762*10^(-6)*exp((29.17*((298.15/T_)-1))+(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p31 => 1.971*10^(6)*exp((30.20*((298.15/T_)-1))-(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p32 => 5.074*10^(-7)*exp((30.20*((298.15/T_)-1))+(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p33 => 2.5*10^(3)*exp((30.20*((298.15/T_)-1))-(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p34 => 4*10^(-4)*exp((30.20*((298.15/T_)-1))+(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p35 => 1.010*10^(-14)*exp((-22.52*((298.15/T_)-1))-(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				#p36 => 9.901*10^(13)*exp((-22.52*((298.15/T_)-1))+(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p37 => ifelse(RH > 0.9300exp(-80.0((1/298.15)-(1/T_))), 4.799*10^(-1)*exp((0.98*((298.15/T_)-1))-(39.75(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #Na2SO4	
-				p38 => 2.084*exp((0.98*((298.15/T_)-1))+(39.75(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p39 => ifelse(RH > 0.7997*exp(-80.0*((1/298.15)-(1/T_))), 1.817*exp((-2.65*((298.15/T_)-1))-(38.57(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #(NH4)2SO4
-				p40 => 5.504*10^(-1)*exp((-2.65*((298.15/T_)-1))+(38.57(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p41 => ifelse(RH > 0.7710*exp(-239.0*((1/298.15)-(1/T_))), 1.086*10^(-16)*exp((-71*((298.15/T_)-1))-(2.4(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),  #NH4Cl
-				#p42 => 9.208*10^(15)*exp((-71*((298.15/T_)-1))+(2.4(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p43 => ifelse(RH > 0.7379*exp(-304.0*((1/298.15)-(1/T_))), 1.197*10^(1)*exp((-8.22*((298.15/T_)-1))-(16.01(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #NaNO3
-				p44 => 8.354*10^(-2)*exp((-8.22*((298.15/T_)-1))+(16.01(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p45 => ifelse(RH > 0.7528*exp(-25.0*((1/298.15)-(1/T_))), 3.766*10^(1)*exp((-1.56*((298.15/T_)-1))-(16.90(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #NaCl
-				p46 => 2.655*10^(-2)*exp((-1.56*((298.15/T_)-1))+(16.90(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p47 => ifelse(RH > 0.5200*exp(45*((1/298.15)-(1/T_))), 2.413*10^(4)*exp((0.79*((298.15/T_)-1))-(14.75(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #NaHSO4
-				p48 => 4.144*10^(-5)*exp((0.79*((298.15/T_)-1))+(14.75(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p49 => ifelse(RH > 0.6183*exp(-852.0*((1/298.15)-(1/T_))), 4.199*10^(-17)*exp((-74.735*((298.15/T_)-1))-(6.025(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #NH4NO3
-				#p50 => 2.382*10^(16)*exp((-74.735*((298.15/T_)-1))+(14.75(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p51 => ifelse(RH > 0.4000*exp(-384.0*((1/298.15)-(1/T_))), 1.383*exp((-2.87*((298.15/T_)-1))-(15.83(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #NH4HSO4
-				p52 => 7.231*10^(-1)*exp((-2.87*((298.15/T_)-1))+(15.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p53 => ifelse(RH > 0.6900*exp(-186.0*((1/298.15)-(1/T_))), 2.972*10^(1)*exp((-5.19*((298.15/T_)-1))-(54.40(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), #(NH4)3H(SO4)2
-				p54 => 3.365*10^(-2)*exp((-5.19*((298.15/T_)-1))+(54.40(1+log((2.718),(298.15/T_))-(298.15/T_))))
-															
-
-	]
-
-	#parammap calculates the equilibrium constant and also utlizes the Deliquescence Relative Humidity (DRH) of sleected salts (if DRH > RH, p[x] = 0)
-
-# ╔═╡ 27b0f338-c071-45c6-9b9d-6f910a36748d
-oprob = ODEProblem(ISORROPIA, u₀map, tspan, parammap)
-
-# ╔═╡ 0f35870d-b175-4642-b1ce-13904395c7ff
-sol = solve(oprob, Tsit5()) #testing the model for input conditions at RH = 0.70(70%)
-
-# ╔═╡ b31ecee6-8bb1-46ae-9360-76fdcf5dfc27
-plot(sol)
-
-# ╔═╡ fa5a4623-3bc1-4776-8233-249bdc4bf73f
-function model(T_,RH,a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p,q,r,s,u,v,w,x,y,z,aa,ab,ac,ad,ae,af,ag,ah,ai,aj,ak,al,am,an)
-	
-	u₀map_ = [CANO32s => a,  CA2aq => b, NO3aq => c,  CACL2s => d, CLaq => e, CASO4s => f, SO42aq => g, H2Og => h, K2SO4s => i, Kaq => j, KHSO4s => k, HSO4aq => l, KNO3s => m, KCLs => n, MGSO4s => o, MG2aq => p, MGNO32s => q, MGCL2s => r, Haq => s, NH3g => u, NH3aq => v, NH4aq => w, OHaq => x, HNO3g => y, HNO3aq => z, HCLg => aa, HCLaq => ab, NA2SO4s => ac, NAaq => ad, NH42SO4s => ae, NH4CLs => af, NANO3s => ag, NO3aq => ah, NACLs => ai, NAHSO4s => aj, NH4NO3s => ak, NH4HSO4s => al, NH43HSO42s => am, H2Oaq => an]
-	
-	parammap_ = [p1 => ifelse(RH > 0.4906*exp(-509.4*((1/298.15)-(1/T_))), 6.067*10^(5)*exp((-11.299*((298.15/T_)-1))-(1+log((2.718),(298.15/T_))-(298.15/T_))), 0),
-				p2 => 1.648*10^(-6)*exp((-11.299*((298.15/T_)-1))+(1+log((2.718),(298.15/T_))-(298.15/T_))),
-				p3 => ifelse(RH > 0.2830*exp(-551.1*((1/298.15)-(1/T_))), 7.974*10^(11)*exp((-14.087*((298.15/T_)-1))-(1+log((2.718),(298.15/T_))-(298.15/T_))), 0),
-				p4 => 1.254^(-12)*exp((-14.087*((298.15/T_)-1))+(1+log((2.718),(298.15/T_))-(298.15/T_))),
-				p5 => ifelse(RH > 0.9700*exp(-((1/298.15)-(1/T_))), 4.319*10^(-5)*exp((((298.15/T_)-1))-(1+log((2.718),(298.15/T_))-(298.15/T_))), 0),
-				p6 => 2.315*10^(4)*exp((((298.15/T_)-1))+(1+log((2.718),(298.15/T_))-(298.15/T_))),
-				p7 => ifelse(RH > 0.9751*exp(-35.6*((1/298.15)-(1/T_))), 1.569*10^(-2)*exp((-9.589*((298.15/T_)-1))-(45.807(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), 
-				p8 => 6.373*10^(-1)*exp((-9.589*((298.15/T_)-1))+(45.807(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p9 => ifelse(RH > 0.8600*exp(-((1/298.15)-(1/T_))), 24.016*exp((-8.423*((298.15/T_)-1))-(17.964(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p10 => 4.164*10^(-2)*exp((-8.423*((298.15/T_)-1))+(17.964(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p11 => ifelse(RH > 0.9248*exp(-((1/298.15)-(1/T_))), 0.872*exp((14.075*((298.15/T_)-1))-(19.388(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), 
-				p12 => 1.147*exp((14.075*((298.15/T_)-1))+(19.388(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p13 => ifelse(RH > 0.8426*exp(-158.9*((1/298.15)-(1/T_))), 8.680*exp((-6.167*((298.15/T_)-1))-(19.953(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p14 => 1.152*10^(-1)*exp((-6.167*((298.15/T_)-1))+(19.953(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p15 => ifelse(RH > 0.8613*exp(714.5*((1/298.15)-(1/T_))), 1.079*10^(5)*exp((36.798*((298.15/T_)-1))-((1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p16 => 9.268*10^(-6)*exp((36.798*((298.15/T_)-1))+((1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p17 => ifelse(RH > 0.5400*exp(-230.2*((1/298.15)-(1/T_))), 2.507*10^(15)*exp((-8.754*((298.15/T_)-1))-((1+log((2.718),(298.15/T_))-(298.15/T_)))), 0), 
-				p18 => 3.989*10^(-16)*exp((-8.754*((298.15/T_)-1))+((1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p19 => ifelse(RH > 0.3284*exp(-42.23*((1/298.15)-(1/T_))), 9.557*10^(21)*exp((-1.347*((298.15/T_)-1))-((1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p20 => 1.046*10^(-22)*exp((-1.347*((298.15/T_)-1))+((1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p21 => 1.015*10^(-2)*exp((8.85*((298.15/T_)-1))-(25.14(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p22 => 9.852*10^(1)*exp((8.85*((298.15/T_)-1))+(25.14(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p23 => 5.764*10^(1)*exp((13.79*((298.15/T_)-1))-(-5.39(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p24 => 1.735*10^(-2)*exp((13.79*((298.15/T_)-1))+(-5.39(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p25 => 1.805*10^(-5)*exp((-1.50*((298.15/T_)-1))-(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p26 => 5.540*10^(4)*exp((-1.50*((298.15/T_)-1))+(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p27 => 2.511*10^(6)*exp((29.17*((298.15/T_)-1))-(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p28 => 3.982*10^(-7)*exp((29.17*((298.15/T_)-1))+(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p29 => 2.1*10^(5)*exp((29.17*((298.15/T_)-1))-(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p30 => 4.762*10^(-6)*exp((29.17*((298.15/T_)-1))+(16.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p31 => 1.971*10^(6)*exp((30.20*((298.15/T_)-1))-(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p32 => 5.074*10^(-7)*exp((30.20*((298.15/T_)-1))+(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p33 => 2.5*10^(3)*exp((30.20*((298.15/T_)-1))-(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p34 => 4*10^(-4)*exp((30.20*((298.15/T_)-1))+(19.91(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p35 => 1.010*10^(-14)*exp((-22.52*((298.15/T_)-1))-(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				#p36 => 9.901*10^(13)*exp((-22.52*((298.15/T_)-1))+(26.92(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p37 => ifelse(RH > 0.9300exp(-80.0((1/298.15)-(1/T_))), 4.799*10^(-1)*exp((0.98*((298.15/T_)-1))-(39.75(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),		
-				p38 => 2.084*exp((0.98*((298.15/T_)-1))+(39.75(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p39 => ifelse(RH > 0.7997*exp(-80.0*((1/298.15)-(1/T_))), 1.817*exp((-2.65*((298.15/T_)-1))-(38.57(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p40 => 5.504*10^(-1)*exp((-2.65*((298.15/T_)-1))+(38.57(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p41 => ifelse(RH > 0.7710*exp(-239.0*((1/298.15)-(1/T_))), 1.086*10^(-16)*exp((-71*((298.15/T_)-1))-(2.4(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				#p42 => 9.208*10^(15)*exp((-71*((298.15/T_)-1))+(2.4(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p43 => ifelse(RH > 0.7379*exp(-304.0*((1/298.15)-(1/T_))), 1.197*10^(1)*exp((-8.22*((298.15/T_)-1))-(16.01(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p44 => 8.354*10^(-2)*exp((-8.22*((298.15/T_)-1))+(16.01(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p45 => ifelse(RH > 0.7528*exp(-25.0*((1/298.15)-(1/T_))), 3.766*10^(1)*exp((-1.56*((298.15/T_)-1))-(16.90(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p46 => 2.655*10^(-2)*exp((-1.56*((298.15/T_)-1))+(16.90(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p47 => ifelse(RH > 0.5200*exp(45*((1/298.15)-(1/T_))), 2.413*10^(4)*exp((0.79*((298.15/T_)-1))-(14.75(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p48 => 4.144*10^(-5)*exp((0.79*((298.15/T_)-1))+(14.75(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p49 => ifelse(RH > 0.6183*exp(-852.0*((1/298.15)-(1/T_))), 4.199*10^(-17)*exp((-74.735*((298.15/T_)-1))-(6.025(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				#p50 => 2.382*10^(16)*exp((-74.735*((298.15/T_)-1))+(14.75(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p51 => ifelse(RH > 0.4000*exp(-384.0*((1/298.15)-(1/T_))), 1.383*exp((-2.87*((298.15/T_)-1))-(15.83(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p52 => 7.231*10^(-1)*exp((-2.87*((298.15/T_)-1))+(15.83(1+log((2.718),(298.15/T_))-(298.15/T_)))),
-				p53 => ifelse(RH > 0.6900*exp(-186.0*((1/298.15)-(1/T_))), 2.972*10^(1)*exp((-5.19*((298.15/T_)-1))-(54.40(1+log((2.718),(298.15/T_))-(298.15/T_)))), 0),
-				p54 => 3.365*10^(-2)*exp((-5.19*((298.15/T_)-1))+(54.40(1+log((2.718),(298.15/T_))-(298.15/T_))))
-												
-
-	]
-	
-	oprob_ = ODEProblem(ISORROPIA, u₀map_, tspan, parammap_)
-	sol_ = solve(oprob_, Tsit5())
-end
-
-#Examples  
-  
-# ╔═╡ 5bbdc0e8-97ce-4706-a266-4e9df3fbe058
-model(298.15,0.75,0.0,0.4,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.33,0.0,10.0,0.0,0.0,0.0,0.0,0.0,0.0,10.0,0.0,3.4,0.0,0.0,2.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0) 
-#testing model for input conditions at RH = 0.75(75%)
-
-# ╔═╡ 9c67e2ba-1d50-4e63-95bc-dd769bfce93e
-model(298.15,0.80,0.0,0.4,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.33,0.0,10.0,0.0,0.0,0.0,0.0,0.0,0.0,10.0,0.0,3.4,0.0,0.0,2.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0) 
-#testing model for input conditions at RH = 0.80(80%)
-
-# ╔═╡ cd76be8b-4267-4996-8fc1-0facfa149f97
-model(298.15,0.85,0.0,0.4,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.33,0.0,10.0,0.0,0.0,0.0,0.0,0.0,0.0,10.0,0.0,3.4,0.0,0.0,2.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0) 
-#testing model for input conditions at RH = 0.85(85%)
-
-# ╔═╡ 5c74962a-4b28-4fab-8fc2-e86c6f533cd7
-model(298.15,0.90,0.0,0.4,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.33,0.0,10.0,0.0,0.0,0.0,0.0,0.0,0.0,10.0,0.0,3.4,0.0,0.0,2.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0) 
-#testing model for input conditions at RH = 0.90(90%)
-
-# ╔═╡ dba4af4a-85ac-457e-ad5c-8d7820a6e942
-model(298.15,0.98,0.0,0.4,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.33,0.0,10.0,0.0,0.0,0.0,0.0,0.0,0.0,10.0,0.0,3.4,0.0,0.0,2.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0) 
-#testing model for input conditions at RH = 0.98(98%)
-
-# ╔═╡ 00000000-0000-0000-0000-000000000001
-PLUTO_PROJECT_TOML_CONTENTS = """
-[deps]
-Catalyst = "479239e8-5488-4da2-87a7-35f2df7eef83"
-DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
-DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
-ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
-NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
-OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
-
-[compat]
-Catalyst = "~10.2.0"
-DiffEqBase = "~6.77.1"
-DifferentialEquations = "~6.20.0"
-ModelingToolkit = "~7.1.3"
-NonlinearSolve = "~0.3.12"
-OrdinaryDiffEq = "~5.71.0"
-Plots = "~1.25.2"
-StochasticDiffEq = "~6.41.0"
-"""
-
-# ╔═╡ 00000000-0000-0000-0000-000000000002
-PLUTO_MANIFEST_TOML_CONTENTS = """
-# This file is machine-generated - editing it directly is not advised
-
-[[AbstractAlgebra]]
-deps = ["GroupsCore", "InteractiveUtils", "LinearAlgebra", "Markdown", "Random", "RandomExtensions", "SparseArrays", "Test"]
-git-tree-sha1 = "b5c1a469f9afd3a9d79930d2cead20c562458aae"
-uuid = "c3fe647b-3220-5bb0-a1ea-a7954cac585d"
-version = "0.22.3"
-
-[[AbstractTrees]]
-git-tree-sha1 = "03e0550477d86222521d254b741d470ba17ea0b5"
-uuid = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
-version = "0.3.4"
-
-[[Adapt]]
-deps = ["LinearAlgebra"]
-git-tree-sha1 = "84918055d15b3114ede17ac6a7182f68870c16f7"
-uuid = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-version = "3.3.1"
-
-[[ArgCheck]]
-git-tree-sha1 = "dedbbb2ddb876f899585c4ec4433265e3017215a"
-uuid = "dce04be8-c92d-5529-be00-80e4d2c0e197"
-version = "2.1.0"
-
-[[ArgTools]]
-uuid = "0dad84c5-d112-42e6-8d28-ef12dabb789f"
-
-[[ArnoldiMethod]]
-deps = ["LinearAlgebra", "Random", "StaticArrays"]
-git-tree-sha1 = "62e51b39331de8911e4a7ff6f5aaf38a5f4cc0ae"
-uuid = "ec485272-7323-5ecc-a04f-4719b315124d"
-version = "0.2.0"
-
-[[ArrayInterface]]
-deps = ["Compat", "IfElse", "LinearAlgebra", "Requires", "SparseArrays", "Static"]
-git-tree-sha1 = "265b06e2b1f6a216e0e8f183d28e4d354eab3220"
-uuid = "4fba245c-0d91-5ea0-9b3e-6abc04ee57a9"
-version = "3.2.1"
-
-[[ArrayLayouts]]
-deps = ["FillArrays", "LinearAlgebra", "SparseArrays"]
-git-tree-sha1 = "e1ba79094cae97b688fb42d31cbbfd63a69706e4"
-uuid = "4c555306-a7a7-4459-81d9-ec55ddd5c99a"
-version = "0.7.8"
-
-[[Artifacts]]
-uuid = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
-
-[[AutoHashEquals]]
-git-tree-sha1 = "45bb6705d93be619b81451bb2006b7ee5d4e4453"
-uuid = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
-version = "0.2.0"
-
-[[BandedMatrices]]
-deps = ["ArrayLayouts", "FillArrays", "LinearAlgebra", "Random", "SparseArrays"]
-git-tree-sha1 = "ce68f8c2162062733f9b4c9e3700d5efc4a8ec47"
-uuid = "aae01518-5342-5314-be14-df237901396f"
-version = "0.16.11"
-
-[[BangBang]]
-deps = ["Compat", "ConstructionBase", "Future", "InitialValues", "LinearAlgebra", "Requires", "Setfield", "Tables", "ZygoteRules"]
-git-tree-sha1 = "0ad226aa72d8671f20d0316e03028f0ba1624307"
-uuid = "198e06fe-97b7-11e9-32a5-e1d131e6ad66"
-version = "0.3.32"
-
-[[Base64]]
-uuid = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
-
-[[Baselet]]
-git-tree-sha1 = "aebf55e6d7795e02ca500a689d326ac979aaf89e"
-uuid = "9718e550-a3fa-408a-8086-8db961cd8217"
-version = "0.1.1"
-
-[[Bijections]]
-git-tree-sha1 = "705e7822597b432ebe152baa844b49f8026df090"
-uuid = "e2ed5e7c-b2de-5872-ae92-c73ca462fb04"
-version = "0.1.3"
-
-[[BitTwiddlingConvenienceFunctions]]
-deps = ["Static"]
-git-tree-sha1 = "bc1317f71de8dce26ea67fcdf7eccc0d0693b75b"
-uuid = "62783981-4cbd-42fc-bca8-16325de8dc4b"
-version = "0.1.1"
-
-[[BoundaryValueDiffEq]]
-deps = ["BandedMatrices", "DiffEqBase", "FiniteDiff", "ForwardDiff", "LinearAlgebra", "NLsolve", "Reexport", "SparseArrays"]
-git-tree-sha1 = "fe34902ac0c3a35d016617ab7032742865756d7d"
-uuid = "764a87c0-6b3e-53db-9096-fe964310641d"
-version = "2.7.1"
-
-[[Bzip2_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "19a35467a82e236ff51bc17a3a44b69ef35185a2"
-uuid = "6e34b625-4abd-537c-b88f-471c36dfa7a0"
-version = "1.0.8+0"
-
-[[CEnum]]
-git-tree-sha1 = "215a9aa4a1f23fbd05b92769fdd62559488d70e9"
-uuid = "fa961155-64e5-5f13-b03f-caf6b980ea82"
-version = "0.4.1"
-
-[[CPUSummary]]
-deps = ["Hwloc", "IfElse", "Static"]
-git-tree-sha1 = "87b0c9c6ee0124d6c1f4ce8cb035dcaf9f90b803"
-uuid = "2a0fbf3d-bb9c-48f3-b0a9-814d99fd7ab9"
-version = "0.1.6"
-
-[[CSTParser]]
-deps = ["Tokenize"]
-git-tree-sha1 = "f9a6389348207faf5e5c62cbc7e89d19688d338a"
-uuid = "00ebfdb7-1f24-5e51-bd34-a7502290713f"
-version = "3.3.0"
-
-[[Cairo_jll]]
-deps = ["Artifacts", "Bzip2_jll", "Fontconfig_jll", "FreeType2_jll", "Glib_jll", "JLLWrappers", "LZO_jll", "Libdl", "Pixman_jll", "Pkg", "Xorg_libXext_jll", "Xorg_libXrender_jll", "Zlib_jll", "libpng_jll"]
-git-tree-sha1 = "f2202b55d816427cd385a9a4f3ffb226bee80f99"
-uuid = "83423d85-b0ee-5818-9007-b63ccbeb887a"
-version = "1.16.1+0"
-
-[[Catalyst]]
-deps = ["AbstractAlgebra", "DataStructures", "DiffEqBase", "DiffEqJump", "DocStringExtensions", "Graphs", "Latexify", "MacroTools", "ModelingToolkit", "Parameters", "Reexport", "Requires", "SparseArrays", "Symbolics"]
-git-tree-sha1 = "f1734a45a0799abc58b08bc1e9bbee75d68a61a2"
-uuid = "479239e8-5488-4da2-87a7-35f2df7eef83"
-version = "10.2.0"
-
-[[ChainRulesCore]]
-deps = ["Compat", "LinearAlgebra", "SparseArrays"]
-git-tree-sha1 = "4c26b4e9e91ca528ea212927326ece5918a04b47"
-uuid = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
-version = "1.11.2"
-
-[[ChangesOfVariables]]
-deps = ["ChainRulesCore", "LinearAlgebra", "Test"]
-git-tree-sha1 = "bf98fa45a0a4cee295de98d4c1462be26345b9a1"
-uuid = "9e997f8a-9a97-42d5-a9f1-ce6bfc15e2c0"
-version = "0.1.2"
-
-[[CloseOpenIntervals]]
-deps = ["ArrayInterface", "Static"]
-git-tree-sha1 = "7b8f09d58294dc8aa13d91a8544b37c8a1dcbc06"
-uuid = "fb6a15b2-703c-40df-9091-08a04967cfa9"
-version = "0.1.4"
-
-[[ColorSchemes]]
-deps = ["ColorTypes", "Colors", "FixedPointNumbers", "Random"]
-git-tree-sha1 = "a851fec56cb73cfdf43762999ec72eff5b86882a"
-uuid = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
-version = "3.15.0"
-
-[[ColorTypes]]
-deps = ["FixedPointNumbers", "Random"]
-git-tree-sha1 = "024fe24d83e4a5bf5fc80501a314ce0d1aa35597"
-uuid = "3da002f7-5984-5a60-b8a6-cbb66c0b333f"
-version = "0.11.0"
-
-[[Colors]]
-deps = ["ColorTypes", "FixedPointNumbers", "Reexport"]
-git-tree-sha1 = "417b0ed7b8b838aa6ca0a87aadf1bb9eb111ce40"
-uuid = "5ae59095-9a9b-59fe-a467-6f913c188581"
-version = "0.12.8"
-
-[[Combinatorics]]
-git-tree-sha1 = "08c8b6831dc00bfea825826be0bc8336fc369860"
-uuid = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
-version = "1.0.2"
-
-[[CommonMark]]
-deps = ["Crayons", "JSON", "URIs"]
-git-tree-sha1 = "7a0d74b8b007c8170dd48166fdc4be049bf68f70"
-uuid = "a80b9123-70ca-4bc0-993e-6e3bcb318db6"
-version = "0.8.4"
-
-[[CommonSolve]]
-git-tree-sha1 = "68a0743f578349ada8bc911a5cbd5a2ef6ed6d1f"
-uuid = "38540f10-b2f7-11e9-35d8-d573e4eb0ff2"
-version = "0.2.0"
-
-[[CommonSubexpressions]]
-deps = ["MacroTools", "Test"]
-git-tree-sha1 = "7b8a93dba8af7e3b42fecabf646260105ac373f7"
-uuid = "bbf7d656-a473-5ed7-a52c-81e309532950"
-version = "0.3.0"
-
-[[Compat]]
-deps = ["Base64", "Dates", "DelimitedFiles", "Distributed", "InteractiveUtils", "LibGit2", "Libdl", "LinearAlgebra", "Markdown", "Mmap", "Pkg", "Printf", "REPL", "Random", "SHA", "Serialization", "SharedArrays", "Sockets", "SparseArrays", "Statistics", "Test", "UUIDs", "Unicode"]
-git-tree-sha1 = "44c37b4636bc54afac5c574d2d02b625349d6582"
-uuid = "34da2185-b29b-5c13-b0c7-acf172513d20"
-version = "3.41.0"
-
-[[CompilerSupportLibraries_jll]]
-deps = ["Artifacts", "Libdl"]
-uuid = "e66e0078-7015-5450-92f7-15fbd957f2ae"
-
-[[CompositeTypes]]
-git-tree-sha1 = "d5b014b216dc891e81fea299638e4c10c657b582"
-uuid = "b152e2b5-7a66-4b01-a709-34e65c35f657"
-version = "0.1.2"
-
-[[CompositionsBase]]
-git-tree-sha1 = "455419f7e328a1a2493cabc6428d79e951349769"
-uuid = "a33af91c-f02d-484b-be07-31d278c5ca2b"
-version = "0.1.1"
-
-[[ConstructionBase]]
-deps = ["LinearAlgebra"]
-git-tree-sha1 = "f74e9d5388b8620b4cee35d4c5a618dd4dc547f4"
-uuid = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
-version = "1.3.0"
-
-[[Contour]]
-deps = ["StaticArrays"]
-git-tree-sha1 = "9f02045d934dc030edad45944ea80dbd1f0ebea7"
-uuid = "d38c429a-6771-53c6-b99e-75d170b6e991"
-version = "0.5.7"
-
-[[Crayons]]
-git-tree-sha1 = "3f71217b538d7aaee0b69ab47d9b7724ca8afa0d"
-uuid = "a8cc5b0e-0ffa-5ad4-8c14-923d3ee1735f"
-version = "4.0.4"
-
-[[DEDataArrays]]
-deps = ["ArrayInterface", "DocStringExtensions", "LinearAlgebra", "RecursiveArrayTools", "SciMLBase", "StaticArrays"]
-git-tree-sha1 = "31186e61936fbbccb41d809ad4338c9f7addf7ae"
-uuid = "754358af-613d-5f8d-9788-280bf1605d4c"
-version = "0.2.0"
-
-[[DataAPI]]
-git-tree-sha1 = "cc70b17275652eb47bc9e5f81635981f13cea5c8"
-uuid = "9a962f9c-6df0-11e9-0e5d-c546b8b5ee8a"
-version = "1.9.0"
-
-[[DataStructures]]
-deps = ["Compat", "InteractiveUtils", "OrderedCollections"]
-git-tree-sha1 = "3daef5523dd2e769dad2365274f760ff5f282c7d"
-uuid = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
-version = "0.18.11"
-
-[[DataValueInterfaces]]
-git-tree-sha1 = "bfc1187b79289637fa0ef6d4436ebdfe6905cbd6"
-uuid = "e2d170a0-9d28-54be-80f0-106bbe20a464"
-version = "1.0.0"
-
-[[Dates]]
-deps = ["Printf"]
-uuid = "ade2ca70-3891-5945-98fb-dc099432e06a"
-
-[[DefineSingletons]]
-git-tree-sha1 = "77b4ca280084423b728662fe040e5ff8819347c5"
-uuid = "244e2a9f-e319-4986-a169-4d1fe445cd52"
-version = "0.1.1"
-
-[[DelayDiffEq]]
-deps = ["ArrayInterface", "DataStructures", "DiffEqBase", "LinearAlgebra", "Logging", "NonlinearSolve", "OrdinaryDiffEq", "Printf", "RecursiveArrayTools", "Reexport", "UnPack"]
-git-tree-sha1 = "9edf8c0f27c16df239d31dd55b154096e4c41cda"
-uuid = "bcd4f6db-9728-5f36-b5f7-82caef46ccdb"
-version = "5.31.2"
-
-[[DelimitedFiles]]
-deps = ["Mmap"]
-uuid = "8bb1440f-4735-579b-a4ab-409b98df4dab"
-
-[[DensityInterface]]
-deps = ["InverseFunctions", "Test"]
-git-tree-sha1 = "80c3e8639e3353e5d2912fb3a1916b8455e2494b"
-uuid = "b429d917-457f-4dbc-8f4c-0cc954292b1d"
-version = "0.4.0"
-
-[[DiffEqBase]]
-deps = ["ArrayInterface", "ChainRulesCore", "DEDataArrays", "DataStructures", "Distributions", "DocStringExtensions", "FastBroadcast", "ForwardDiff", "FunctionWrappers", "IterativeSolvers", "LabelledArrays", "LinearAlgebra", "Logging", "MuladdMacro", "NonlinearSolve", "Parameters", "PreallocationTools", "Printf", "RecursiveArrayTools", "RecursiveFactorization", "Reexport", "Requires", "SciMLBase", "Setfield", "SparseArrays", "StaticArrays", "Statistics", "SuiteSparse", "ZygoteRules"]
-git-tree-sha1 = "976a9d34dff2f2bad494cec64c32d54168db21b2"
-uuid = "2b5f629d-d688-5b77-993f-72d75c75574e"
-version = "6.77.1"
-
-[[DiffEqCallbacks]]
-deps = ["DataStructures", "DiffEqBase", "ForwardDiff", "LinearAlgebra", "NLsolve", "OrdinaryDiffEq", "Parameters", "RecipesBase", "RecursiveArrayTools", "SciMLBase", "StaticArrays"]
-git-tree-sha1 = "a615f494f0c10d0a21f895aa0b65b2056c37d17b"
-uuid = "459566f4-90b8-5000-8ac3-15dfb0a30def"
-version = "2.18.0"
-
-[[DiffEqFinancial]]
-deps = ["DiffEqBase", "DiffEqNoiseProcess", "LinearAlgebra", "Markdown", "RandomNumbers"]
-git-tree-sha1 = "db08e0def560f204167c58fd0637298e13f58f73"
-uuid = "5a0ffddc-d203-54b0-88ba-2c03c0fc2e67"
-version = "2.4.0"
-
-[[DiffEqJump]]
-deps = ["ArrayInterface", "Compat", "DataStructures", "DiffEqBase", "FunctionWrappers", "Graphs", "LinearAlgebra", "PoissonRandom", "Random", "RandomNumbers", "RecursiveArrayTools", "Reexport", "StaticArrays", "TreeViews", "UnPack"]
-git-tree-sha1 = "0aa2d003ec9efe2a93f93ae722de05a870ffc0b2"
-uuid = "c894b116-72e5-5b58-be3c-e6d8d4ac2b12"
-version = "8.0.0"
-
-[[DiffEqNoiseProcess]]
-deps = ["DiffEqBase", "Distributions", "LinearAlgebra", "Optim", "PoissonRandom", "QuadGK", "Random", "Random123", "RandomNumbers", "RecipesBase", "RecursiveArrayTools", "Requires", "ResettableStacks", "SciMLBase", "StaticArrays", "Statistics"]
-git-tree-sha1 = "d6839a44a268c69ef0ed927b22a6f43c8a4c2e73"
-uuid = "77a26b50-5914-5dd7-bc55-306e6241c503"
-version = "5.9.0"
-
-[[DiffEqPhysics]]
-deps = ["DiffEqBase", "DiffEqCallbacks", "ForwardDiff", "LinearAlgebra", "Printf", "Random", "RecipesBase", "RecursiveArrayTools", "Reexport", "StaticArrays"]
-git-tree-sha1 = "8f23c6f36f6a6eb2cbd6950e28ec7c4b99d0e4c9"
-uuid = "055956cb-9e8b-5191-98cc-73ae4a59e68a"
-version = "3.9.0"
-
-[[DiffResults]]
-deps = ["StaticArrays"]
-git-tree-sha1 = "c18e98cba888c6c25d1c3b048e4b3380ca956805"
-uuid = "163ba53b-c6d8-5494-b064-1a9d43ac40c5"
-version = "1.0.3"
-
-[[DiffRules]]
-deps = ["LogExpFunctions", "NaNMath", "Random", "SpecialFunctions"]
-git-tree-sha1 = "9bc5dac3c8b6706b58ad5ce24cffd9861f07c94f"
-uuid = "b552c78f-8df3-52c6-915a-8e097449b14b"
-version = "1.9.0"
-
-[[DifferentialEquations]]
-deps = ["BoundaryValueDiffEq", "DelayDiffEq", "DiffEqBase", "DiffEqCallbacks", "DiffEqFinancial", "DiffEqJump", "DiffEqNoiseProcess", "DiffEqPhysics", "DimensionalPlotRecipes", "LinearAlgebra", "MultiScaleArrays", "OrdinaryDiffEq", "ParameterizedFunctions", "Random", "RecursiveArrayTools", "Reexport", "SteadyStateDiffEq", "StochasticDiffEq", "Sundials"]
-git-tree-sha1 = "91df208ee040be7960c408d4681bf91974bcb4f4"
-uuid = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
-version = "6.20.0"
-
-[[DimensionalPlotRecipes]]
-deps = ["LinearAlgebra", "RecipesBase"]
-git-tree-sha1 = "af883a26bbe6e3f5f778cb4e1b81578b534c32a6"
-uuid = "c619ae07-58cd-5f6d-b883-8f17bd6a98f9"
-version = "1.2.0"
-
-[[Distances]]
-deps = ["LinearAlgebra", "SparseArrays", "Statistics", "StatsAPI"]
-git-tree-sha1 = "3258d0659f812acde79e8a74b11f17ac06d0ca04"
-uuid = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
-version = "0.10.7"
-
-[[Distributed]]
-deps = ["Random", "Serialization", "Sockets"]
-uuid = "8ba89e20-285c-5b6f-9357-94700520ee1b"
-
-[[Distributions]]
-deps = ["ChainRulesCore", "DensityInterface", "FillArrays", "LinearAlgebra", "PDMats", "Printf", "QuadGK", "Random", "SparseArrays", "SpecialFunctions", "Statistics", "StatsBase", "StatsFuns", "Test"]
-git-tree-sha1 = "c1724611e6ae29c6094c8d9850e3136297ba7fff"
-uuid = "31c24e10-a181-5473-b8eb-7969acd0382f"
-version = "0.25.36"
-
-[[DocStringExtensions]]
-deps = ["LibGit2"]
-git-tree-sha1 = "b19534d1895d702889b219c382a6e18010797f0b"
-uuid = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-version = "0.8.6"
-
-[[DomainSets]]
-deps = ["CompositeTypes", "IntervalSets", "LinearAlgebra", "StaticArrays", "Statistics"]
-git-tree-sha1 = "5f5f0b750ac576bcf2ab1d7782959894b304923e"
-uuid = "5b8099bc-c8ec-5219-889f-1d9e522a28bf"
-version = "0.5.9"
-
-[[Downloads]]
-deps = ["ArgTools", "LibCURL", "NetworkOptions"]
-uuid = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
-
-[[DynamicPolynomials]]
-deps = ["DataStructures", "Future", "LinearAlgebra", "MultivariatePolynomials", "MutableArithmetics", "Pkg", "Reexport", "Test"]
-git-tree-sha1 = "585de0d658506cf0fe5808026edff662bef5bf03"
-uuid = "7c1d4256-1411-5781-91ec-d7bc3513ac07"
-version = "0.4.1"
-
-[[EarCut_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "3f3a2501fa7236e9b911e0f7a588c657e822bb6d"
-uuid = "5ae413db-bbd1-5e63-b57d-d24a61df00f5"
-version = "2.2.3+0"
-
-[[EllipsisNotation]]
-deps = ["ArrayInterface"]
-git-tree-sha1 = "3fe985505b4b667e1ae303c9ca64d181f09d5c05"
-uuid = "da5c29d0-fa7d-589e-88eb-ea29b0a81949"
-version = "1.1.3"
-
-[[Expat_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "b3bfd02e98aedfa5cf885665493c5598c350cd2f"
-uuid = "2e619515-83b5-522b-bb60-26c02a35a201"
-version = "2.2.10+0"
-
-[[ExponentialUtilities]]
-deps = ["ArrayInterface", "LinearAlgebra", "Printf", "Requires", "SparseArrays"]
-git-tree-sha1 = "1b873816d2cfc8c0fcb1edcb08e67fdf630a70b7"
-uuid = "d4d017d3-3776-5f7e-afef-a10c40355c18"
-version = "1.10.2"
-
-[[ExprTools]]
-git-tree-sha1 = "b7e3d17636b348f005f11040025ae8c6f645fe92"
-uuid = "e2ba6199-217a-4e67-a87a-7c52f15ade04"
-version = "0.1.6"
-
-[[FFMPEG]]
-deps = ["FFMPEG_jll"]
-git-tree-sha1 = "b57e3acbe22f8484b4b5ff66a7499717fe1a9cc8"
-uuid = "c87230d0-a227-11e9-1b43-d7ebe4e7570a"
-version = "0.4.1"
-
-[[FFMPEG_jll]]
-deps = ["Artifacts", "Bzip2_jll", "FreeType2_jll", "FriBidi_jll", "JLLWrappers", "LAME_jll", "Libdl", "Ogg_jll", "OpenSSL_jll", "Opus_jll", "Pkg", "Zlib_jll", "libass_jll", "libfdk_aac_jll", "libvorbis_jll", "x264_jll", "x265_jll"]
-git-tree-sha1 = "d8a578692e3077ac998b50c0217dfd67f21d1e5f"
-uuid = "b22a6f82-2f65-5046-a5b2-351ab43fb4e5"
-version = "4.4.0+0"
-
-[[FastBroadcast]]
-deps = ["LinearAlgebra", "Polyester", "Static"]
-git-tree-sha1 = "e32a81c505ab234c992ca978f31ed8b0dabbc327"
-uuid = "7034ab61-46d4-4ed7-9d0f-46aef9175898"
-version = "0.1.11"
-
-[[FastClosures]]
-git-tree-sha1 = "acebe244d53ee1b461970f8910c235b259e772ef"
-uuid = "9aa1b823-49e4-5ca5-8b0f-3971ec8bab6a"
-version = "0.3.2"
-
-[[FillArrays]]
-deps = ["LinearAlgebra", "Random", "SparseArrays", "Statistics"]
-git-tree-sha1 = "8756f9935b7ccc9064c6eef0bff0ad643df733a3"
-uuid = "1a297f60-69ca-5386-bcde-b61e274b549b"
-version = "0.12.7"
-
-[[FiniteDiff]]
-deps = ["ArrayInterface", "LinearAlgebra", "Requires", "SparseArrays", "StaticArrays"]
-git-tree-sha1 = "8b3c09b56acaf3c0e581c66638b85c8650ee9dca"
-uuid = "6a86dc24-6348-571c-b903-95158fe2bd41"
-version = "2.8.1"
-
-[[FixedPointNumbers]]
-deps = ["Statistics"]
-git-tree-sha1 = "335bfdceacc84c5cdf16aadc768aa5ddfc5383cc"
-uuid = "53c48c17-4a7d-5ca2-90c5-79b7896eea93"
-version = "0.8.4"
-
-[[Fontconfig_jll]]
-deps = ["Artifacts", "Bzip2_jll", "Expat_jll", "FreeType2_jll", "JLLWrappers", "Libdl", "Libuuid_jll", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "21efd19106a55620a188615da6d3d06cd7f6ee03"
-uuid = "a3f928ae-7b40-5064-980b-68af3947d34b"
-version = "2.13.93+0"
-
-[[Formatting]]
-deps = ["Printf"]
-git-tree-sha1 = "8339d61043228fdd3eb658d86c926cb282ae72a8"
-uuid = "59287772-0a20-5a39-b81b-1366585eb4c0"
-version = "0.4.2"
-
-[[ForwardDiff]]
-deps = ["CommonSubexpressions", "DiffResults", "DiffRules", "LinearAlgebra", "LogExpFunctions", "NaNMath", "Preferences", "Printf", "Random", "SpecialFunctions", "StaticArrays"]
-git-tree-sha1 = "2b72a5624e289ee18256111657663721d59c143e"
-uuid = "f6369f11-7733-5829-9624-2563aa707210"
-version = "0.10.24"
-
-[[FreeType2_jll]]
-deps = ["Artifacts", "Bzip2_jll", "JLLWrappers", "Libdl", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "87eb71354d8ec1a96d4a7636bd57a7347dde3ef9"
-uuid = "d7e528f0-a631-5988-bf34-fe36492bcfd7"
-version = "2.10.4+0"
-
-[[FriBidi_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "aa31987c2ba8704e23c6c8ba8a4f769d5d7e4f91"
-uuid = "559328eb-81f9-559d-9380-de523a88c83c"
-version = "1.0.10+0"
-
-[[FunctionWrappers]]
-git-tree-sha1 = "241552bc2209f0fa068b6415b1942cc0aa486bcc"
-uuid = "069b7b12-0de2-55c6-9aab-29f3d0a68a2e"
-version = "1.1.2"
-
-[[Future]]
-deps = ["Random"]
-uuid = "9fa8497b-333b-5362-9e8d-4d0656e87820"
-
-[[GLFW_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Libglvnd_jll", "Pkg", "Xorg_libXcursor_jll", "Xorg_libXi_jll", "Xorg_libXinerama_jll", "Xorg_libXrandr_jll"]
-git-tree-sha1 = "0c603255764a1fa0b61752d2bec14cfbd18f7fe8"
-uuid = "0656b61e-2033-5cc2-a64a-77c0f6c09b89"
-version = "3.3.5+1"
-
-[[GR]]
-deps = ["Base64", "DelimitedFiles", "GR_jll", "HTTP", "JSON", "Libdl", "LinearAlgebra", "Pkg", "Printf", "Random", "Serialization", "Sockets", "Test", "UUIDs"]
-git-tree-sha1 = "30f2b340c2fff8410d89bfcdc9c0a6dd661ac5f7"
-uuid = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
-version = "0.62.1"
-
-[[GR_jll]]
-deps = ["Artifacts", "Bzip2_jll", "Cairo_jll", "FFMPEG_jll", "Fontconfig_jll", "GLFW_jll", "JLLWrappers", "JpegTurbo_jll", "Libdl", "Libtiff_jll", "Pixman_jll", "Pkg", "Qt5Base_jll", "Zlib_jll", "libpng_jll"]
-git-tree-sha1 = "fd75fa3a2080109a2c0ec9864a6e14c60cca3866"
-uuid = "d2c73de3-f751-5644-a686-071e5b155ba9"
-version = "0.62.0+0"
-
-[[GeometryBasics]]
-deps = ["EarCut_jll", "IterTools", "LinearAlgebra", "StaticArrays", "StructArrays", "Tables"]
-git-tree-sha1 = "58bcdf5ebc057b085e58d95c138725628dd7453c"
-uuid = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
-version = "0.4.1"
-
-[[Gettext_jll]]
-deps = ["Artifacts", "CompilerSupportLibraries_jll", "JLLWrappers", "Libdl", "Libiconv_jll", "Pkg", "XML2_jll"]
-git-tree-sha1 = "9b02998aba7bf074d14de89f9d37ca24a1a0b046"
-uuid = "78b55507-aeef-58d4-861c-77aaff3498b1"
-version = "0.21.0+0"
-
-[[Glib_jll]]
-deps = ["Artifacts", "Gettext_jll", "JLLWrappers", "Libdl", "Libffi_jll", "Libiconv_jll", "Libmount_jll", "PCRE_jll", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "74ef6288d071f58033d54fd6708d4bc23a8b8972"
-uuid = "7746bdde-850d-59dc-9ae8-88ece973131d"
-version = "2.68.3+1"
-
-[[Graphite2_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "344bf40dcab1073aca04aa0df4fb092f920e4011"
-uuid = "3b182d85-2403-5c21-9c21-1e1f0cc25472"
-version = "1.3.14+0"
-
-[[Graphs]]
-deps = ["ArnoldiMethod", "DataStructures", "Distributed", "Inflate", "LinearAlgebra", "Random", "SharedArrays", "SimpleTraits", "SparseArrays", "Statistics"]
-git-tree-sha1 = "92243c07e786ea3458532e199eb3feee0e7e08eb"
-uuid = "86223c79-3864-5bf0-83f7-82e725a168b6"
-version = "1.4.1"
-
-[[Grisu]]
-git-tree-sha1 = "53bb909d1151e57e2484c3d1b53e19552b887fb2"
-uuid = "42e2da0e-8278-4e71-bc24-59509adca0fe"
-version = "1.0.2"
-
-[[GroupsCore]]
-deps = ["Markdown", "Random"]
-git-tree-sha1 = "9e1a5e9f3b81ad6a5c613d181664a0efc6fe6dd7"
-uuid = "d5909c97-4eac-4ecc-a3dc-fdd0858a4120"
-version = "0.4.0"
-
-[[HTTP]]
-deps = ["Base64", "Dates", "IniFile", "Logging", "MbedTLS", "NetworkOptions", "Sockets", "URIs"]
-git-tree-sha1 = "0fa77022fe4b511826b39c894c90daf5fce3334a"
-uuid = "cd3eb016-35fb-5094-929b-558a96fad6f3"
-version = "0.9.17"
-
-[[HarfBuzz_jll]]
-deps = ["Artifacts", "Cairo_jll", "Fontconfig_jll", "FreeType2_jll", "Glib_jll", "Graphite2_jll", "JLLWrappers", "Libdl", "Libffi_jll", "Pkg"]
-git-tree-sha1 = "129acf094d168394e80ee1dc4bc06ec835e510a3"
-uuid = "2e76f6c2-a576-52d4-95c1-20adfe4de566"
-version = "2.8.1+1"
-
-[[HostCPUFeatures]]
-deps = ["BitTwiddlingConvenienceFunctions", "IfElse", "Libdl", "Static"]
-git-tree-sha1 = "8f0dc80088981ab55702b04bba38097a44a1a3a9"
-uuid = "3e5b6fbb-0976-4d2c-9146-d79de83f2fb0"
-version = "0.1.5"
-
-[[Hwloc]]
-deps = ["Hwloc_jll"]
-git-tree-sha1 = "92d99146066c5c6888d5a3abc871e6a214388b91"
-uuid = "0e44f5e4-bd66-52a0-8798-143a42290a1d"
-version = "2.0.0"
-
-[[Hwloc_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "3395d4d4aeb3c9d31f5929d32760d8baeee88aaf"
-uuid = "e33a78d0-f292-5ffc-b300-72abe9b543c8"
-version = "2.5.0+0"
-
-[[IfElse]]
-git-tree-sha1 = "debdd00ffef04665ccbb3e150747a77560e8fad1"
-uuid = "615f187c-cbe4-4ef1-ba3b-2fcf58d6d173"
-version = "0.1.1"
-
-[[Inflate]]
-git-tree-sha1 = "f5fc07d4e706b84f72d54eedcc1c13d92fb0871c"
-uuid = "d25df0c9-e2be-5dd7-82c8-3ad0b3e990b9"
-version = "0.1.2"
-
-[[IniFile]]
-deps = ["Test"]
-git-tree-sha1 = "098e4d2c533924c921f9f9847274f2ad89e018b8"
-uuid = "83e8ac13-25f8-5344-8a64-a9f2b223428f"
-version = "0.5.0"
-
-[[InitialValues]]
-git-tree-sha1 = "7f6a4508b4a6f46db5ccd9799a3fc71ef5cad6e6"
-uuid = "22cec73e-a1b8-11e9-2c92-598750a2cf9c"
-version = "0.2.11"
-
-[[InteractiveUtils]]
-deps = ["Markdown"]
-uuid = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
-
-[[IntervalSets]]
-deps = ["Dates", "EllipsisNotation", "Statistics"]
-git-tree-sha1 = "3cc368af3f110a767ac786560045dceddfc16758"
-uuid = "8197267c-284f-5f27-9208-e0e47529a953"
-version = "0.5.3"
-
-[[InverseFunctions]]
-deps = ["Test"]
-git-tree-sha1 = "a7254c0acd8e62f1ac75ad24d5db43f5f19f3c65"
-uuid = "3587e190-3f89-42d0-90ee-14403ec27112"
-version = "0.1.2"
-
-[[IrrationalConstants]]
-git-tree-sha1 = "7fd44fd4ff43fc60815f8e764c0f352b83c49151"
-uuid = "92d709cd-6900-40b7-9082-c6be49f344b6"
-version = "0.1.1"
-
-[[IterTools]]
-git-tree-sha1 = "fa6287a4469f5e048d763df38279ee729fbd44e5"
-uuid = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
-version = "1.4.0"
-
-[[IterativeSolvers]]
-deps = ["LinearAlgebra", "Printf", "Random", "RecipesBase", "SparseArrays"]
-git-tree-sha1 = "1169632f425f79429f245113b775a0e3d121457c"
-uuid = "42fd0dbc-a981-5370-80f2-aaf504508153"
-version = "0.9.2"
-
-[[IteratorInterfaceExtensions]]
-git-tree-sha1 = "a3f24677c21f5bbe9d2a714f95dcd58337fb2856"
-uuid = "82899510-4779-5014-852e-03e436cf321d"
-version = "1.0.0"
-
-[[JLLWrappers]]
-deps = ["Preferences"]
-git-tree-sha1 = "642a199af8b68253517b80bd3bfd17eb4e84df6e"
-uuid = "692b3bcd-3c85-4b1f-b108-f13ce0eb3210"
-version = "1.3.0"
-
-[[JSON]]
-deps = ["Dates", "Mmap", "Parsers", "Unicode"]
-git-tree-sha1 = "8076680b162ada2a031f707ac7b4953e30667a37"
-uuid = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
-version = "0.21.2"
-
-[[JpegTurbo_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "d735490ac75c5cb9f1b00d8b5509c11984dc6943"
-uuid = "aacddb02-875f-59d6-b918-886e6ef4fbf8"
-version = "2.1.0+0"
-
-[[JuliaFormatter]]
-deps = ["CSTParser", "CommonMark", "DataStructures", "Pkg", "Tokenize"]
-git-tree-sha1 = "e1a91af84efba5228480bd20aeef1dd902d553cf"
-uuid = "98e50ef6-434e-11e9-1051-2b60c6c9e899"
-version = "0.19.2"
-
-[[LAME_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "f6250b16881adf048549549fba48b1161acdac8c"
-uuid = "c1c5ebd0-6772-5130-a774-d5fcae4a789d"
-version = "3.100.1+0"
-
-[[LZO_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "e5b909bcf985c5e2605737d2ce278ed791b89be6"
-uuid = "dd4b983a-f0e5-5f8d-a1b7-129d4a5fb1ac"
-version = "2.10.1+0"
-
-[[LaTeXStrings]]
-git-tree-sha1 = "f2355693d6778a178ade15952b7ac47a4ff97996"
-uuid = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
-version = "1.3.0"
-
-[[LabelledArrays]]
-deps = ["ArrayInterface", "ChainRulesCore", "LinearAlgebra", "MacroTools", "StaticArrays"]
-git-tree-sha1 = "3609bbf5feba7b22fb35fe7cb207c8c8d2e2fc5b"
-uuid = "2ee39098-c373-598a-b85f-a56591580800"
-version = "1.6.7"
-
-[[Latexify]]
-deps = ["Formatting", "InteractiveUtils", "LaTeXStrings", "MacroTools", "Markdown", "Printf", "Requires"]
-git-tree-sha1 = "a8f4f279b6fa3c3c4f1adadd78a621b13a506bce"
-uuid = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
-version = "0.15.9"
-
-[[LayoutPointers]]
-deps = ["ArrayInterface", "LinearAlgebra", "ManualMemory", "SIMDTypes", "Static"]
-git-tree-sha1 = "83b56449c39342a47f3fcdb3bc782bd6d66e1d97"
-uuid = "10f19ff3-798f-405d-979b-55457f8fc047"
-version = "0.1.4"
-
-[[LibCURL]]
-deps = ["LibCURL_jll", "MozillaCACerts_jll"]
-uuid = "b27032c2-a3e7-50c8-80cd-2d36dbcbfd21"
-
-[[LibCURL_jll]]
-deps = ["Artifacts", "LibSSH2_jll", "Libdl", "MbedTLS_jll", "Zlib_jll", "nghttp2_jll"]
-uuid = "deac9b47-8bc7-5906-a0fe-35ac56dc84c0"
-
-[[LibGit2]]
-deps = ["Base64", "NetworkOptions", "Printf", "SHA"]
-uuid = "76f85450-5226-5b5a-8eaa-529ad045b433"
-
-[[LibSSH2_jll]]
-deps = ["Artifacts", "Libdl", "MbedTLS_jll"]
-uuid = "29816b5a-b9ab-546f-933c-edad1886dfa8"
-
-[[Libdl]]
-uuid = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
-
-[[Libffi_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "0b4a5d71f3e5200a7dff793393e09dfc2d874290"
-uuid = "e9f186c6-92d2-5b65-8a66-fee21dc1b490"
-version = "3.2.2+1"
-
-[[Libgcrypt_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Libgpg_error_jll", "Pkg"]
-git-tree-sha1 = "64613c82a59c120435c067c2b809fc61cf5166ae"
-uuid = "d4300ac3-e22c-5743-9152-c294e39db1e4"
-version = "1.8.7+0"
-
-[[Libglvnd_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libX11_jll", "Xorg_libXext_jll"]
-git-tree-sha1 = "7739f837d6447403596a75d19ed01fd08d6f56bf"
-uuid = "7e76a0d4-f3c7-5321-8279-8d96eeed0f29"
-version = "1.3.0+3"
-
-[[Libgpg_error_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "c333716e46366857753e273ce6a69ee0945a6db9"
-uuid = "7add5ba3-2f88-524e-9cd5-f83b8a55f7b8"
-version = "1.42.0+0"
-
-[[Libiconv_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "42b62845d70a619f063a7da093d995ec8e15e778"
-uuid = "94ce4f54-9a6c-5748-9c1c-f9c7231a4531"
-version = "1.16.1+1"
-
-[[Libmount_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "9c30530bf0effd46e15e0fdcf2b8636e78cbbd73"
-uuid = "4b2f31a3-9ecc-558c-b454-b3730dcb73e9"
-version = "2.35.0+0"
-
-[[Libtiff_jll]]
-deps = ["Artifacts", "JLLWrappers", "JpegTurbo_jll", "Libdl", "Pkg", "Zlib_jll", "Zstd_jll"]
-git-tree-sha1 = "340e257aada13f95f98ee352d316c3bed37c8ab9"
-uuid = "89763e89-9b03-5906-acba-b20f662cd828"
-version = "4.3.0+0"
-
-[[Libuuid_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "7f3efec06033682db852f8b3bc3c1d2b0a0ab066"
-uuid = "38a345b3-de98-5d2b-a5d3-14cd9215e700"
-version = "2.36.0+0"
-
-[[LineSearches]]
-deps = ["LinearAlgebra", "NLSolversBase", "NaNMath", "Parameters", "Printf"]
-git-tree-sha1 = "f27132e551e959b3667d8c93eae90973225032dd"
-uuid = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
-version = "7.1.1"
-
-[[LinearAlgebra]]
-deps = ["Libdl"]
-uuid = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-
-[[LogExpFunctions]]
-deps = ["ChainRulesCore", "ChangesOfVariables", "DocStringExtensions", "InverseFunctions", "IrrationalConstants", "LinearAlgebra"]
-git-tree-sha1 = "be9eef9f9d78cecb6f262f3c10da151a6c5ab827"
-uuid = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
-version = "0.3.5"
-
-[[Logging]]
-uuid = "56ddb016-857b-54e1-b83d-db4d58db5568"
-
-[[LoopVectorization]]
-deps = ["ArrayInterface", "CPUSummary", "CloseOpenIntervals", "DocStringExtensions", "HostCPUFeatures", "IfElse", "LayoutPointers", "LinearAlgebra", "OffsetArrays", "PolyesterWeave", "Requires", "SIMDDualNumbers", "SLEEFPirates", "Static", "ThreadingUtilities", "UnPack", "VectorizationBase"]
-git-tree-sha1 = "9e10579c154f785b911d9ceb96c33fcc1a661171"
-uuid = "bdcacae8-1622-11e9-2a5c-532679323890"
-version = "0.12.99"
-
-[[MacroTools]]
-deps = ["Markdown", "Random"]
-git-tree-sha1 = "3d3e902b31198a27340d0bf00d6ac452866021cf"
-uuid = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
-version = "0.5.9"
-
-[[ManualMemory]]
-git-tree-sha1 = "9cb207b18148b2199db259adfa923b45593fe08e"
-uuid = "d125e4d3-2237-4719-b19c-fa641b8a4667"
-version = "0.1.6"
-
-[[Markdown]]
-deps = ["Base64"]
-uuid = "d6f4376e-aef5-505a-96c1-9c027394607a"
-
-[[MbedTLS]]
-deps = ["Dates", "MbedTLS_jll", "Random", "Sockets"]
-git-tree-sha1 = "1c38e51c3d08ef2278062ebceade0e46cefc96fe"
-uuid = "739be429-bea8-5141-9913-cc70e7f3736d"
-version = "1.0.3"
-
-[[MbedTLS_jll]]
-deps = ["Artifacts", "Libdl"]
-uuid = "c8ffd9c3-330d-5841-b78e-0817d7145fa1"
-
-[[Measures]]
-git-tree-sha1 = "e498ddeee6f9fdb4551ce855a46f54dbd900245f"
-uuid = "442fdcdd-2543-5da2-b0f3-8c86c306513e"
-version = "0.3.1"
-
-[[Metatheory]]
-deps = ["AutoHashEquals", "DataStructures", "Dates", "DocStringExtensions", "Parameters", "Reexport", "TermInterface", "ThreadsX", "TimerOutputs"]
-git-tree-sha1 = "0d3b2feb3168e4deb78361d3b5bb5c2e51ea5271"
-uuid = "e9d8d322-4543-424a-9be4-0cc815abe26c"
-version = "1.3.2"
-
-[[MicroCollections]]
-deps = ["BangBang", "Setfield"]
-git-tree-sha1 = "4f65bdbbe93475f6ff9ea6969b21532f88d359be"
-uuid = "128add7d-3638-4c79-886c-908ea0c25c34"
-version = "0.1.1"
-
-[[Missings]]
-deps = ["DataAPI"]
-git-tree-sha1 = "bf210ce90b6c9eed32d25dbcae1ebc565df2687f"
-uuid = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
-version = "1.0.2"
-
-[[Mmap]]
-uuid = "a63ad114-7e13-5084-954f-fe012c677804"
-
-[[ModelingToolkit]]
-deps = ["AbstractTrees", "ArrayInterface", "ConstructionBase", "DataStructures", "DiffEqBase", "DiffEqCallbacks", "DiffEqJump", "DiffRules", "Distributed", "Distributions", "DocStringExtensions", "DomainSets", "Graphs", "IfElse", "InteractiveUtils", "JuliaFormatter", "LabelledArrays", "Latexify", "Libdl", "LinearAlgebra", "MacroTools", "NaNMath", "NonlinearSolve", "RecursiveArrayTools", "Reexport", "Requires", "RuntimeGeneratedFunctions", "SafeTestsets", "SciMLBase", "Serialization", "Setfield", "SparseArrays", "SpecialFunctions", "StaticArrays", "SymbolicUtils", "Symbolics", "UnPack", "Unitful"]
-git-tree-sha1 = "d9bbe2b0141a2387dcb6cd00a7019a5b39d7d8b9"
-uuid = "961ee093-0014-501f-94e3-6117800e7a78"
-version = "7.1.3"
-
-[[MozillaCACerts_jll]]
-uuid = "14a3606d-f60d-562e-9121-12d972cd8159"
-
-[[MuladdMacro]]
-git-tree-sha1 = "c6190f9a7fc5d9d5915ab29f2134421b12d24a68"
-uuid = "46d2c3a1-f734-5fdb-9937-b9b9aeba4221"
-version = "0.2.2"
-
-[[MultiScaleArrays]]
-deps = ["DiffEqBase", "FiniteDiff", "ForwardDiff", "LinearAlgebra", "OrdinaryDiffEq", "Random", "RecursiveArrayTools", "SparseDiffTools", "Statistics", "StochasticDiffEq", "TreeViews"]
-git-tree-sha1 = "258f3be6770fe77be8870727ba9803e236c685b8"
-uuid = "f9640e96-87f6-5992-9c3b-0743c6a49ffa"
-version = "1.8.1"
-
-[[MultivariatePolynomials]]
-deps = ["DataStructures", "LinearAlgebra", "MutableArithmetics"]
-git-tree-sha1 = "fa6ce8c91445e7cd54de662064090b14b1089a6d"
-uuid = "102ac46a-7ee4-5c85-9060-abc95bfdeaa3"
-version = "0.4.2"
-
-[[MutableArithmetics]]
-deps = ["LinearAlgebra", "SparseArrays", "Test"]
-git-tree-sha1 = "7bb6853d9afec54019c1397c6eb610b9b9a19525"
-uuid = "d8a4904e-b15c-11e9-3269-09a3773c0cb0"
-version = "0.3.1"
-
-[[NLSolversBase]]
-deps = ["DiffResults", "Distributed", "FiniteDiff", "ForwardDiff"]
-git-tree-sha1 = "50310f934e55e5ca3912fb941dec199b49ca9b68"
-uuid = "d41bc354-129a-5804-8e4c-c37616107c6c"
-version = "7.8.2"
-
-[[NLsolve]]
-deps = ["Distances", "LineSearches", "LinearAlgebra", "NLSolversBase", "Printf", "Reexport"]
-git-tree-sha1 = "019f12e9a1a7880459d0173c182e6a99365d7ac1"
-uuid = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
-version = "4.5.1"
-
-[[NaNMath]]
-git-tree-sha1 = "bfe47e760d60b82b66b61d2d44128b62e3a369fb"
-uuid = "77ba4419-2d1f-58cd-9bb1-8ffee604a2e3"
-version = "0.3.5"
-
-[[NetworkOptions]]
-uuid = "ca575930-c2e3-43a9-ace4-1e988b2c1908"
-
-[[NonlinearSolve]]
-deps = ["ArrayInterface", "FiniteDiff", "ForwardDiff", "IterativeSolvers", "LinearAlgebra", "RecursiveArrayTools", "RecursiveFactorization", "Reexport", "SciMLBase", "Setfield", "StaticArrays", "UnPack"]
-git-tree-sha1 = "8dc3be3e9edf976a3e79363b3bd2ad776a627c31"
-uuid = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
-version = "0.3.12"
-
-[[OffsetArrays]]
-deps = ["Adapt"]
-git-tree-sha1 = "043017e0bdeff61cfbb7afeb558ab29536bbb5ed"
-uuid = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
-version = "1.10.8"
-
-[[Ogg_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "7937eda4681660b4d6aeeecc2f7e1c81c8ee4e2f"
-uuid = "e7412a2a-1a6e-54c0-be00-318e2571c051"
-version = "1.3.5+0"
-
-[[OpenBLAS_jll]]
-deps = ["Artifacts", "CompilerSupportLibraries_jll", "Libdl"]
-uuid = "4536629a-c528-5b80-bd46-f80d51c5b363"
-
-[[OpenLibm_jll]]
-deps = ["Artifacts", "Libdl"]
-uuid = "05823500-19ac-5b8b-9628-191a04bc5112"
-
-[[OpenSSL_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "15003dcb7d8db3c6c857fda14891a539a8f2705a"
-uuid = "458c3c95-2e84-50aa-8efc-19380b2a3a95"
-version = "1.1.10+0"
-
-[[OpenSpecFun_jll]]
-deps = ["Artifacts", "CompilerSupportLibraries_jll", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "13652491f6856acfd2db29360e1bbcd4565d04f1"
-uuid = "efe28fd5-8261-553b-a9e1-b2916fc3738e"
-version = "0.5.5+0"
-
-[[Optim]]
-deps = ["Compat", "FillArrays", "ForwardDiff", "LineSearches", "LinearAlgebra", "NLSolversBase", "NaNMath", "Parameters", "PositiveFactorizations", "Printf", "SparseArrays", "StatsBase"]
-git-tree-sha1 = "35d435b512fbab1d1a29138b5229279925eba369"
-uuid = "429524aa-4258-5aef-a3af-852621145aeb"
-version = "1.5.0"
-
-[[Opus_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "51a08fb14ec28da2ec7a927c4337e4332c2a4720"
-uuid = "91d4177d-7536-5919-b921-800302f37372"
-version = "1.3.2+0"
-
-[[OrderedCollections]]
-git-tree-sha1 = "85f8e6578bf1f9ee0d11e7bb1b1456435479d47c"
-uuid = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
-version = "1.4.1"
-
-[[OrdinaryDiffEq]]
-deps = ["Adapt", "ArrayInterface", "DataStructures", "DiffEqBase", "DocStringExtensions", "ExponentialUtilities", "FastClosures", "FiniteDiff", "ForwardDiff", "LinearAlgebra", "Logging", "LoopVectorization", "MacroTools", "MuladdMacro", "NLsolve", "Polyester", "PreallocationTools", "RecursiveArrayTools", "Reexport", "SparseArrays", "SparseDiffTools", "StaticArrays", "UnPack"]
-git-tree-sha1 = "aab86872f5dccb1a6251ae7935b286e784dadba0"
-uuid = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
-version = "5.71.0"
-
-[[PCRE_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "b2a7af664e098055a7529ad1a900ded962bca488"
-uuid = "2f80f16e-611a-54ab-bc61-aa92de5b98fc"
-version = "8.44.0+0"
-
-[[PDMats]]
-deps = ["LinearAlgebra", "SparseArrays", "SuiteSparse"]
-git-tree-sha1 = "ee26b350276c51697c9c2d88a072b339f9f03d73"
-uuid = "90014a1f-27ba-587c-ab20-58faa44d9150"
-version = "0.11.5"
-
-[[ParameterizedFunctions]]
-deps = ["DataStructures", "DiffEqBase", "DocStringExtensions", "Latexify", "LinearAlgebra", "ModelingToolkit", "Reexport", "SciMLBase"]
-git-tree-sha1 = "3baa1ad75b77f406988be4dc0364e01cf16127e7"
-uuid = "65888b18-ceab-5e60-b2b9-181511a3b968"
-version = "5.12.2"
-
-[[Parameters]]
-deps = ["OrderedCollections", "UnPack"]
-git-tree-sha1 = "34c0e9ad262e5f7fc75b10a9952ca7692cfc5fbe"
-uuid = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-version = "0.12.3"
-
-[[Parsers]]
-deps = ["Dates"]
-git-tree-sha1 = "ae4bbcadb2906ccc085cf52ac286dc1377dceccc"
-uuid = "69de0a69-1ddd-5017-9359-2bf0b02dc9f0"
-version = "2.1.2"
-
-[[Pixman_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "b4f5d02549a10e20780a24fce72bea96b6329e29"
-uuid = "30392449-352a-5448-841d-b1acce4e97dc"
-version = "0.40.1+0"
-
-[[Pkg]]
-deps = ["Artifacts", "Dates", "Downloads", "LibGit2", "Libdl", "Logging", "Markdown", "Printf", "REPL", "Random", "SHA", "Serialization", "TOML", "Tar", "UUIDs", "p7zip_jll"]
-uuid = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-
-[[PlotThemes]]
-deps = ["PlotUtils", "Requires", "Statistics"]
-git-tree-sha1 = "a3a964ce9dc7898193536002a6dd892b1b5a6f1d"
-uuid = "ccf2f8ad-2431-5c83-bf29-c5338b663b6a"
-version = "2.0.1"
-
-[[PlotUtils]]
-deps = ["ColorSchemes", "Colors", "Dates", "Printf", "Random", "Reexport", "Statistics"]
-git-tree-sha1 = "e4fe0b50af3130ddd25e793b471cb43d5279e3e6"
-uuid = "995b91a9-d308-5afd-9ec6-746e21dbc043"
-version = "1.1.1"
-
-[[Plots]]
-deps = ["Base64", "Contour", "Dates", "Downloads", "FFMPEG", "FixedPointNumbers", "GR", "GeometryBasics", "JSON", "Latexify", "LinearAlgebra", "Measures", "NaNMath", "PlotThemes", "PlotUtils", "Printf", "REPL", "Random", "RecipesBase", "RecipesPipeline", "Reexport", "Requires", "Scratch", "Showoff", "SparseArrays", "Statistics", "StatsBase", "UUIDs", "UnicodeFun"]
-git-tree-sha1 = "65ebc27d8c00c84276f14aaf4ff63cbe12016c70"
-uuid = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-version = "1.25.2"
-
-[[PoissonRandom]]
-deps = ["Random", "Statistics", "Test"]
-git-tree-sha1 = "44d018211a56626288b5d3f8c6497d28c26dc850"
-uuid = "e409e4f3-bfea-5376-8464-e040bb5c01ab"
-version = "0.4.0"
-
-[[Polyester]]
-deps = ["ArrayInterface", "BitTwiddlingConvenienceFunctions", "CPUSummary", "IfElse", "ManualMemory", "PolyesterWeave", "Requires", "Static", "StrideArraysCore", "ThreadingUtilities"]
-git-tree-sha1 = "892b8d9dd3c7987a4d0fd320f0a421dd90b5d09d"
-uuid = "f517fe37-dbe3-4b94-8317-1923a5111588"
-version = "0.5.4"
-
-[[PolyesterWeave]]
-deps = ["BitTwiddlingConvenienceFunctions", "CPUSummary", "IfElse", "Static", "ThreadingUtilities"]
-git-tree-sha1 = "a3ff99bf561183ee20386aec98ab8f4a12dc724a"
-uuid = "1d0040c9-8b98-4ee7-8388-3f51789ca0ad"
-version = "0.1.2"
-
-[[PositiveFactorizations]]
-deps = ["LinearAlgebra"]
-git-tree-sha1 = "17275485f373e6673f7e7f97051f703ed5b15b20"
-uuid = "85a6dd25-e78a-55b7-8502-1745935b8125"
-version = "0.2.4"
-
-[[PreallocationTools]]
-deps = ["Adapt", "ArrayInterface", "ForwardDiff", "LabelledArrays"]
-git-tree-sha1 = "435379f01c1e6f7ca65cf46fdd403226f1d36e37"
-uuid = "d236fae5-4411-538c-8e31-a6e3d9e00b46"
-version = "0.2.1"
-
-[[Preferences]]
-deps = ["TOML"]
-git-tree-sha1 = "00cfd92944ca9c760982747e9a1d0d5d86ab1e5a"
-uuid = "21216c6a-2e73-6563-6e65-726566657250"
-version = "1.2.2"
-
-[[Printf]]
-deps = ["Unicode"]
-uuid = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-
-[[Qt5Base_jll]]
-deps = ["Artifacts", "CompilerSupportLibraries_jll", "Fontconfig_jll", "Glib_jll", "JLLWrappers", "Libdl", "Libglvnd_jll", "OpenSSL_jll", "Pkg", "Xorg_libXext_jll", "Xorg_libxcb_jll", "Xorg_xcb_util_image_jll", "Xorg_xcb_util_keysyms_jll", "Xorg_xcb_util_renderutil_jll", "Xorg_xcb_util_wm_jll", "Zlib_jll", "xkbcommon_jll"]
-git-tree-sha1 = "ad368663a5e20dbb8d6dc2fddeefe4dae0781ae8"
-uuid = "ea2cea3b-5b76-57ae-a6ef-0a8af62496e1"
-version = "5.15.3+0"
-
-[[QuadGK]]
-deps = ["DataStructures", "LinearAlgebra"]
-git-tree-sha1 = "78aadffb3efd2155af139781b8a8df1ef279ea39"
-uuid = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
-version = "2.4.2"
-
-[[REPL]]
-deps = ["InteractiveUtils", "Markdown", "Sockets", "Unicode"]
-uuid = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
-
-[[Random]]
-deps = ["Serialization"]
-uuid = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-
-[[Random123]]
-deps = ["Libdl", "Random", "RandomNumbers"]
-git-tree-sha1 = "0e8b146557ad1c6deb1367655e052276690e71a3"
-uuid = "74087812-796a-5b5d-8853-05524746bad3"
-version = "1.4.2"
-
-[[RandomExtensions]]
-deps = ["Random", "SparseArrays"]
-git-tree-sha1 = "062986376ce6d394b23d5d90f01d81426113a3c9"
-uuid = "fb686558-2515-59ef-acaa-46db3789a887"
-version = "0.4.3"
-
-[[RandomNumbers]]
-deps = ["Random", "Requires"]
-git-tree-sha1 = "043da614cc7e95c703498a491e2c21f58a2b8111"
-uuid = "e6cf234a-135c-5ec9-84dd-332b85af5143"
-version = "1.5.3"
-
-[[RecipesBase]]
-git-tree-sha1 = "6bf3f380ff52ce0832ddd3a2a7b9538ed1bcca7d"
-uuid = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
-version = "1.2.1"
-
-[[RecipesPipeline]]
-deps = ["Dates", "NaNMath", "PlotUtils", "RecipesBase"]
-git-tree-sha1 = "7ad0dfa8d03b7bcf8c597f59f5292801730c55b8"
-uuid = "01d81517-befc-4cb6-b9ec-a95719d0359c"
-version = "0.4.1"
-
-[[RecursiveArrayTools]]
-deps = ["ArrayInterface", "ChainRulesCore", "DocStringExtensions", "FillArrays", "LinearAlgebra", "RecipesBase", "Requires", "StaticArrays", "Statistics", "ZygoteRules"]
-git-tree-sha1 = "c944fa4adbb47be43376359811c0a14757bdc8a8"
-uuid = "731186ca-8d62-57ce-b412-fbd966d074cd"
-version = "2.20.0"
-
-[[RecursiveFactorization]]
-deps = ["LinearAlgebra", "LoopVectorization", "Polyester", "StrideArraysCore", "TriangularSolve"]
-git-tree-sha1 = "b7edd69c796b30985ea6dfeda8504cdb7cf77e9f"
-uuid = "f2c3362d-daeb-58d1-803e-2bc74f2840b4"
-version = "0.2.5"
-
-[[Reexport]]
-git-tree-sha1 = "45e428421666073eab6f2da5c9d310d99bb12f9b"
-uuid = "189a3867-3050-52da-a836-e630ba90ab69"
-version = "1.2.2"
-
-[[Referenceables]]
-deps = ["Adapt"]
-git-tree-sha1 = "e681d3bfa49cd46c3c161505caddf20f0e62aaa9"
-uuid = "42d2dcc6-99eb-4e98-b66c-637b7d73030e"
-version = "0.1.2"
-
-[[Requires]]
-deps = ["UUIDs"]
-git-tree-sha1 = "8f82019e525f4d5c669692772a6f4b0a58b06a6a"
-uuid = "ae029012-a4dd-5104-9daa-d747884805df"
-version = "1.2.0"
-
-[[ResettableStacks]]
-deps = ["StaticArrays"]
-git-tree-sha1 = "256eeeec186fa7f26f2801732774ccf277f05db9"
-uuid = "ae5879a3-cd67-5da8-be7f-38c6eb64a37b"
-version = "1.1.1"
-
-[[Rmath]]
-deps = ["Random", "Rmath_jll"]
-git-tree-sha1 = "bf3188feca147ce108c76ad82c2792c57abe7b1f"
-uuid = "79098fc4-a85e-5d69-aa6a-4863f24498fa"
-version = "0.7.0"
-
-[[Rmath_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "68db32dff12bb6127bac73c209881191bf0efbb7"
-uuid = "f50d1b31-88e8-58de-be2c-1cc44531875f"
-version = "0.3.0+0"
-
-[[RuntimeGeneratedFunctions]]
-deps = ["ExprTools", "SHA", "Serialization"]
-git-tree-sha1 = "cdc1e4278e91a6ad530770ebb327f9ed83cf10c4"
-uuid = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
-version = "0.5.3"
-
-[[SHA]]
-uuid = "ea8e919c-243c-51af-8825-aaa63cd721ce"
-
-[[SIMDDualNumbers]]
-deps = ["ForwardDiff", "IfElse", "SLEEFPirates", "VectorizationBase"]
-git-tree-sha1 = "62c2da6eb66de8bb88081d20528647140d4daa0e"
-uuid = "3cdde19b-5bb0-4aaf-8931-af3e248e098b"
-version = "0.1.0"
-
-[[SIMDTypes]]
-git-tree-sha1 = "330289636fb8107c5f32088d2741e9fd7a061a5c"
-uuid = "94e857df-77ce-4151-89e5-788b33177be4"
-version = "0.1.0"
-
-[[SLEEFPirates]]
-deps = ["IfElse", "Static", "VectorizationBase"]
-git-tree-sha1 = "1410aad1c6b35862573c01b96cd1f6dbe3979994"
-uuid = "476501e8-09a2-5ece-8869-fb82de89a1fa"
-version = "0.6.28"
-
-[[SafeTestsets]]
-deps = ["Test"]
-git-tree-sha1 = "36ebc5622c82eb9324005cc75e7e2cc51181d181"
-uuid = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
-version = "0.0.1"
-
-[[SciMLBase]]
-deps = ["ArrayInterface", "CommonSolve", "ConstructionBase", "Distributed", "DocStringExtensions", "IteratorInterfaceExtensions", "LinearAlgebra", "Logging", "RecipesBase", "RecursiveArrayTools", "StaticArrays", "Statistics", "Tables", "TreeViews"]
-git-tree-sha1 = "aed443e3bc01ade4555a79bb22f3694e9828c9d4"
-uuid = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
-version = "1.22.1"
-
-[[Scratch]]
-deps = ["Dates"]
-git-tree-sha1 = "0b4b7f1393cff97c33891da2a0bf69c6ed241fda"
-uuid = "6c6a2e73-6563-6170-7368-637461726353"
-version = "1.1.0"
-
-[[Serialization]]
-uuid = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
-
-[[Setfield]]
-deps = ["ConstructionBase", "Future", "MacroTools", "Requires"]
-git-tree-sha1 = "0afd9e6c623e379f593da01f20590bacc26d1d14"
-uuid = "efcf1570-3423-57d1-acb7-fd33fddbac46"
-version = "0.8.1"
-
-[[SharedArrays]]
-deps = ["Distributed", "Mmap", "Random", "Serialization"]
-uuid = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
-
-[[Showoff]]
-deps = ["Dates", "Grisu"]
-git-tree-sha1 = "91eddf657aca81df9ae6ceb20b959ae5653ad1de"
-uuid = "992d4aef-0814-514b-bc4d-f2e9a6c4116f"
-version = "1.0.3"
-
-[[SimpleTraits]]
-deps = ["InteractiveUtils", "MacroTools"]
-git-tree-sha1 = "5d7e3f4e11935503d3ecaf7186eac40602e7d231"
-uuid = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
-version = "0.9.4"
-
-[[Sockets]]
-uuid = "6462fe0b-24de-5631-8697-dd941f90decc"
-
-[[SortingAlgorithms]]
-deps = ["DataStructures"]
-git-tree-sha1 = "b3363d7460f7d098ca0912c69b082f75625d7508"
-uuid = "a2af1166-a08f-5f64-846c-94a0d3cef48c"
-version = "1.0.1"
-
-[[SparseArrays]]
-deps = ["LinearAlgebra", "Random"]
-uuid = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-
-[[SparseDiffTools]]
-deps = ["Adapt", "ArrayInterface", "Compat", "DataStructures", "FiniteDiff", "ForwardDiff", "Graphs", "LinearAlgebra", "Requires", "SparseArrays", "StaticArrays", "VertexSafeGraphs"]
-git-tree-sha1 = "f33a0f6721b270cdf417f0c986e93c973e5913c8"
-uuid = "47a9eef4-7e08-11e9-0b38-333d64bd3804"
-version = "1.19.1"
-
-[[SpecialFunctions]]
-deps = ["ChainRulesCore", "IrrationalConstants", "LogExpFunctions", "OpenLibm_jll", "OpenSpecFun_jll"]
-git-tree-sha1 = "f0bccf98e16759818ffc5d97ac3ebf87eb950150"
-uuid = "276daf66-3868-5448-9aa4-cd146d93841b"
-version = "1.8.1"
-
-[[SplittablesBase]]
-deps = ["Setfield", "Test"]
-git-tree-sha1 = "39c9f91521de844bad65049efd4f9223e7ed43f9"
-uuid = "171d559e-b47b-412a-8079-5efa626c420e"
-version = "0.1.14"
-
-[[Static]]
-deps = ["IfElse"]
-git-tree-sha1 = "7f5a513baec6f122401abfc8e9c074fdac54f6c1"
-uuid = "aedffcd0-7271-4cad-89d0-dc628f76c6d3"
-version = "0.4.1"
-
-[[StaticArrays]]
-deps = ["LinearAlgebra", "Random", "Statistics"]
-git-tree-sha1 = "3c76dde64d03699e074ac02eb2e8ba8254d428da"
-uuid = "90137ffa-7385-5640-81b9-e52037218182"
-version = "1.2.13"
-
-[[Statistics]]
-deps = ["LinearAlgebra", "SparseArrays"]
-uuid = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-
-[[StatsAPI]]
-git-tree-sha1 = "0f2aa8e32d511f758a2ce49208181f7733a0936a"
-uuid = "82ae8749-77ed-4fe6-ae5f-f523153014b0"
-version = "1.1.0"
-
-[[StatsBase]]
-deps = ["DataAPI", "DataStructures", "LinearAlgebra", "LogExpFunctions", "Missings", "Printf", "Random", "SortingAlgorithms", "SparseArrays", "Statistics", "StatsAPI"]
-git-tree-sha1 = "2bb0cb32026a66037360606510fca5984ccc6b75"
-uuid = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
-version = "0.33.13"
-
-[[StatsFuns]]
-deps = ["ChainRulesCore", "InverseFunctions", "IrrationalConstants", "LogExpFunctions", "Reexport", "Rmath", "SpecialFunctions"]
-git-tree-sha1 = "bedb3e17cc1d94ce0e6e66d3afa47157978ba404"
-uuid = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
-version = "0.9.14"
-
-[[SteadyStateDiffEq]]
-deps = ["DiffEqBase", "DiffEqCallbacks", "LinearAlgebra", "NLsolve", "Reexport", "SciMLBase"]
-git-tree-sha1 = "3e057e1f9f12d18cac32011aed9e61eef6c1c0ce"
-uuid = "9672c7b4-1e72-59bd-8a11-6ac3964bc41f"
-version = "1.6.6"
-
-[[StochasticDiffEq]]
-deps = ["Adapt", "ArrayInterface", "DataStructures", "DiffEqBase", "DiffEqJump", "DiffEqNoiseProcess", "DocStringExtensions", "FillArrays", "FiniteDiff", "ForwardDiff", "LinearAlgebra", "Logging", "MuladdMacro", "NLsolve", "OrdinaryDiffEq", "Random", "RandomNumbers", "RecursiveArrayTools", "Reexport", "SparseArrays", "SparseDiffTools", "StaticArrays", "UnPack"]
-git-tree-sha1 = "d6756d0c66aecd5d57ad9d305d7c2526fb5922d9"
-uuid = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
-version = "6.41.0"
-
-[[StrideArraysCore]]
-deps = ["ArrayInterface", "CloseOpenIntervals", "IfElse", "LayoutPointers", "ManualMemory", "Requires", "SIMDTypes", "Static", "ThreadingUtilities"]
-git-tree-sha1 = "12cf3253ebd8e2a3214ae171fbfe51e7e8d8ad28"
-uuid = "7792a7ef-975c-4747-a70f-980b88e8d1da"
-version = "0.2.9"
-
-[[StructArrays]]
-deps = ["Adapt", "DataAPI", "StaticArrays", "Tables"]
-git-tree-sha1 = "2ce41e0d042c60ecd131e9fb7154a3bfadbf50d3"
-uuid = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
-version = "0.6.3"
-
-[[SuiteSparse]]
-deps = ["Libdl", "LinearAlgebra", "Serialization", "SparseArrays"]
-uuid = "4607b0f0-06f3-5cda-b6b1-a6196a1729e9"
-
-[[SuiteSparse_jll]]
-deps = ["Artifacts", "Libdl", "OpenBLAS_jll"]
-uuid = "bea87d4a-7f5b-5778-9afe-8cc45184846c"
-
-[[Sundials]]
-deps = ["CEnum", "DataStructures", "DiffEqBase", "Libdl", "LinearAlgebra", "Logging", "Reexport", "SparseArrays", "Sundials_jll"]
-git-tree-sha1 = "12d529a67c232bd27e9868fbcfad4997435786a5"
-uuid = "c3572dad-4567-51f8-b174-8c6c989267f4"
-version = "4.6.0"
-
-[[Sundials_jll]]
-deps = ["CompilerSupportLibraries_jll", "Libdl", "OpenBLAS_jll", "Pkg", "SuiteSparse_jll"]
-git-tree-sha1 = "013ff4504fc1d475aa80c63b455b6b3a58767db2"
-uuid = "fb77eaff-e24c-56d4-86b1-d163f2edb164"
-version = "5.2.0+1"
-
-[[SymbolicUtils]]
-deps = ["AbstractTrees", "Bijections", "ChainRulesCore", "Combinatorics", "ConstructionBase", "DataStructures", "DocStringExtensions", "DynamicPolynomials", "IfElse", "LabelledArrays", "LinearAlgebra", "Metatheory", "MultivariatePolynomials", "NaNMath", "Setfield", "SparseArrays", "SpecialFunctions", "StaticArrays", "TermInterface", "TimerOutputs"]
-git-tree-sha1 = "1b4d3f3bc8ecc80552b3ee24c7c5155905913931"
-uuid = "d1185830-fcd6-423d-90d6-eec64667417b"
-version = "0.18.2"
-
-[[Symbolics]]
-deps = ["ConstructionBase", "DataStructures", "DiffRules", "Distributions", "DocStringExtensions", "DomainSets", "IfElse", "Latexify", "Libdl", "LinearAlgebra", "MacroTools", "Metatheory", "NaNMath", "RecipesBase", "Reexport", "Requires", "RuntimeGeneratedFunctions", "SciMLBase", "Setfield", "SparseArrays", "SpecialFunctions", "StaticArrays", "SymbolicUtils", "TermInterface", "TreeViews"]
-git-tree-sha1 = "c354713d0e64aa527ffd7298ce1106459540209b"
-uuid = "0c5d862f-8b57-4792-8d23-62f2024744c7"
-version = "4.2.0"
-
-[[TOML]]
-deps = ["Dates"]
-uuid = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
-
-[[TableTraits]]
-deps = ["IteratorInterfaceExtensions"]
-git-tree-sha1 = "c06b2f539df1c6efa794486abfb6ed2022561a39"
-uuid = "3783bdb8-4a98-5b6b-af9a-565f29a5fe9c"
-version = "1.0.1"
-
-[[Tables]]
-deps = ["DataAPI", "DataValueInterfaces", "IteratorInterfaceExtensions", "LinearAlgebra", "TableTraits", "Test"]
-git-tree-sha1 = "bb1064c9a84c52e277f1096cf41434b675cd368b"
-uuid = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-version = "1.6.1"
-
-[[Tar]]
-deps = ["ArgTools", "SHA"]
-uuid = "a4e569a6-e804-4fa4-b0f3-eef7a1d5b13e"
-
-[[TermInterface]]
-git-tree-sha1 = "7aa601f12708243987b88d1b453541a75e3d8c7a"
-uuid = "8ea1fca8-c5ef-4a55-8b96-4e9afe9c9a3c"
-version = "0.2.3"
-
-[[Test]]
-deps = ["InteractiveUtils", "Logging", "Random", "Serialization"]
-uuid = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[[ThreadingUtilities]]
-deps = ["ManualMemory"]
-git-tree-sha1 = "03013c6ae7f1824131b2ae2fc1d49793b51e8394"
-uuid = "8290d209-cae3-49c0-8002-c8c24d57dab5"
-version = "0.4.6"
-
-[[ThreadsX]]
-deps = ["ArgCheck", "BangBang", "ConstructionBase", "InitialValues", "MicroCollections", "Referenceables", "Setfield", "SplittablesBase", "Transducers"]
-git-tree-sha1 = "abcff3ac31c7894550566be533b512f8b059104f"
-uuid = "ac1d9e8a-700a-412c-b207-f0111f4b6c0d"
-version = "0.1.8"
-
-[[TimerOutputs]]
-deps = ["ExprTools", "Printf"]
-git-tree-sha1 = "7cb456f358e8f9d102a8b25e8dfedf58fa5689bc"
-uuid = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
-version = "0.5.13"
-
-[[Tokenize]]
-git-tree-sha1 = "0952c9cee34988092d73a5708780b3917166a0dd"
-uuid = "0796e94c-ce3b-5d07-9a54-7f471281c624"
-version = "0.5.21"
-
-[[Transducers]]
-deps = ["Adapt", "ArgCheck", "BangBang", "Baselet", "CompositionsBase", "DefineSingletons", "Distributed", "InitialValues", "Logging", "Markdown", "MicroCollections", "Requires", "Setfield", "SplittablesBase", "Tables"]
-git-tree-sha1 = "bccb153150744d476a6a8d4facf5299325d5a442"
-uuid = "28d57a85-8fef-5791-bfe6-a80928e7c999"
-version = "0.4.67"
-
-[[TreeViews]]
-deps = ["Test"]
-git-tree-sha1 = "8d0d7a3fe2f30d6a7f833a5f19f7c7a5b396eae6"
-uuid = "a2a6695c-b41b-5b7d-aed9-dbfdeacea5d7"
-version = "0.3.0"
-
-[[TriangularSolve]]
-deps = ["CloseOpenIntervals", "IfElse", "LayoutPointers", "LinearAlgebra", "LoopVectorization", "Polyester", "Static", "VectorizationBase"]
-git-tree-sha1 = "ec9a310324dd2c546c07f33a599ded9c1d00a420"
-uuid = "d5829a12-d9aa-46ab-831f-fb7c9ab06edf"
-version = "0.1.8"
-
-[[URIs]]
-git-tree-sha1 = "97bbe755a53fe859669cd907f2d96aee8d2c1355"
-uuid = "5c2747f8-b7ea-4ff2-ba2e-563bfd36b1d4"
-version = "1.3.0"
-
-[[UUIDs]]
-deps = ["Random", "SHA"]
-uuid = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
-
-[[UnPack]]
-git-tree-sha1 = "387c1f73762231e86e0c9c5443ce3b4a0a9a0c2b"
-uuid = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
-version = "1.0.2"
-
-[[Unicode]]
-uuid = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
-
-[[UnicodeFun]]
-deps = ["REPL"]
-git-tree-sha1 = "53915e50200959667e78a92a418594b428dffddf"
-uuid = "1cfade01-22cf-5700-b092-accc4b62d6e1"
-version = "0.4.1"
-
-[[Unitful]]
-deps = ["ConstructionBase", "Dates", "LinearAlgebra", "Random"]
-git-tree-sha1 = "0992ed0c3ef66b0390e5752fe60054e5ff93b908"
-uuid = "1986cc42-f94f-5a68-af5c-568840ba703d"
-version = "1.9.2"
-
-[[VectorizationBase]]
-deps = ["ArrayInterface", "CPUSummary", "HostCPUFeatures", "Hwloc", "IfElse", "LayoutPointers", "Libdl", "LinearAlgebra", "SIMDTypes", "Static"]
-git-tree-sha1 = "17e5847bb36730d90801170ecd0ce4041a3dde86"
-uuid = "3d5dd08c-fd9d-11e8-17fa-ed2836048c2f"
-version = "0.21.22"
-
-[[VertexSafeGraphs]]
-deps = ["Graphs"]
-git-tree-sha1 = "8351f8d73d7e880bfc042a8b6922684ebeafb35c"
-uuid = "19fa3120-7c27-5ec5-8db8-b0b0aa330d6f"
-version = "0.2.0"
-
-[[Wayland_jll]]
-deps = ["Artifacts", "Expat_jll", "JLLWrappers", "Libdl", "Libffi_jll", "Pkg", "XML2_jll"]
-git-tree-sha1 = "3e61f0b86f90dacb0bc0e73a0c5a83f6a8636e23"
-uuid = "a2964d1f-97da-50d4-b82a-358c7fce9d89"
-version = "1.19.0+0"
-
-[[Wayland_protocols_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "66d72dc6fcc86352f01676e8f0f698562e60510f"
-uuid = "2381bf8a-dfd0-557d-9999-79630e7b1b91"
-version = "1.23.0+0"
-
-[[XML2_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Libiconv_jll", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "1acf5bdf07aa0907e0a37d3718bb88d4b687b74a"
-uuid = "02c8fc9c-b97f-50b9-bbe4-9be30ff0a78a"
-version = "2.9.12+0"
-
-[[XSLT_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Libgcrypt_jll", "Libgpg_error_jll", "Libiconv_jll", "Pkg", "XML2_jll", "Zlib_jll"]
-git-tree-sha1 = "91844873c4085240b95e795f692c4cec4d805f8a"
-uuid = "aed1982a-8fda-507f-9586-7b0439959a61"
-version = "1.1.34+0"
-
-[[Xorg_libX11_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libxcb_jll", "Xorg_xtrans_jll"]
-git-tree-sha1 = "5be649d550f3f4b95308bf0183b82e2582876527"
-uuid = "4f6342f7-b3d2-589e-9d20-edeb45f2b2bc"
-version = "1.6.9+4"
-
-[[Xorg_libXau_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "4e490d5c960c314f33885790ed410ff3a94ce67e"
-uuid = "0c0b7dd1-d40b-584c-a123-a41640f87eec"
-version = "1.0.9+4"
-
-[[Xorg_libXcursor_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libXfixes_jll", "Xorg_libXrender_jll"]
-git-tree-sha1 = "12e0eb3bc634fa2080c1c37fccf56f7c22989afd"
-uuid = "935fb764-8cf2-53bf-bb30-45bb1f8bf724"
-version = "1.2.0+4"
-
-[[Xorg_libXdmcp_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "4fe47bd2247248125c428978740e18a681372dd4"
-uuid = "a3789734-cfe1-5b06-b2d0-1dd0d9d62d05"
-version = "1.1.3+4"
-
-[[Xorg_libXext_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libX11_jll"]
-git-tree-sha1 = "b7c0aa8c376b31e4852b360222848637f481f8c3"
-uuid = "1082639a-0dae-5f34-9b06-72781eeb8cb3"
-version = "1.3.4+4"
-
-[[Xorg_libXfixes_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libX11_jll"]
-git-tree-sha1 = "0e0dc7431e7a0587559f9294aeec269471c991a4"
-uuid = "d091e8ba-531a-589c-9de9-94069b037ed8"
-version = "5.0.3+4"
-
-[[Xorg_libXi_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libXext_jll", "Xorg_libXfixes_jll"]
-git-tree-sha1 = "89b52bc2160aadc84d707093930ef0bffa641246"
-uuid = "a51aa0fd-4e3c-5386-b890-e753decda492"
-version = "1.7.10+4"
-
-[[Xorg_libXinerama_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libXext_jll"]
-git-tree-sha1 = "26be8b1c342929259317d8b9f7b53bf2bb73b123"
-uuid = "d1454406-59df-5ea1-beac-c340f2130bc3"
-version = "1.1.4+4"
-
-[[Xorg_libXrandr_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libXext_jll", "Xorg_libXrender_jll"]
-git-tree-sha1 = "34cea83cb726fb58f325887bf0612c6b3fb17631"
-uuid = "ec84b674-ba8e-5d96-8ba1-2a689ba10484"
-version = "1.5.2+4"
-
-[[Xorg_libXrender_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libX11_jll"]
-git-tree-sha1 = "19560f30fd49f4d4efbe7002a1037f8c43d43b96"
-uuid = "ea2f1a96-1ddc-540d-b46f-429655e07cfa"
-version = "0.9.10+4"
-
-[[Xorg_libpthread_stubs_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "6783737e45d3c59a4a4c4091f5f88cdcf0908cbb"
-uuid = "14d82f49-176c-5ed1-bb49-ad3f5cbd8c74"
-version = "0.1.0+3"
-
-[[Xorg_libxcb_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "XSLT_jll", "Xorg_libXau_jll", "Xorg_libXdmcp_jll", "Xorg_libpthread_stubs_jll"]
-git-tree-sha1 = "daf17f441228e7a3833846cd048892861cff16d6"
-uuid = "c7cfdc94-dc32-55de-ac96-5a1b8d977c5b"
-version = "1.13.0+3"
-
-[[Xorg_libxkbfile_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libX11_jll"]
-git-tree-sha1 = "926af861744212db0eb001d9e40b5d16292080b2"
-uuid = "cc61e674-0454-545c-8b26-ed2c68acab7a"
-version = "1.1.0+4"
-
-[[Xorg_xcb_util_image_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_xcb_util_jll"]
-git-tree-sha1 = "0fab0a40349ba1cba2c1da699243396ff8e94b97"
-uuid = "12413925-8142-5f55-bb0e-6d7ca50bb09b"
-version = "0.4.0+1"
-
-[[Xorg_xcb_util_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libxcb_jll"]
-git-tree-sha1 = "e7fd7b2881fa2eaa72717420894d3938177862d1"
-uuid = "2def613f-5ad1-5310-b15b-b15d46f528f5"
-version = "0.4.0+1"
-
-[[Xorg_xcb_util_keysyms_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_xcb_util_jll"]
-git-tree-sha1 = "d1151e2c45a544f32441a567d1690e701ec89b00"
-uuid = "975044d2-76e6-5fbe-bf08-97ce7c6574c7"
-version = "0.4.0+1"
-
-[[Xorg_xcb_util_renderutil_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_xcb_util_jll"]
-git-tree-sha1 = "dfd7a8f38d4613b6a575253b3174dd991ca6183e"
-uuid = "0d47668e-0667-5a69-a72c-f761630bfb7e"
-version = "0.3.9+1"
-
-[[Xorg_xcb_util_wm_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_xcb_util_jll"]
-git-tree-sha1 = "e78d10aab01a4a154142c5006ed44fd9e8e31b67"
-uuid = "c22f9ab0-d5fe-5066-847c-f4bb1cd4e361"
-version = "0.4.1+1"
-
-[[Xorg_xkbcomp_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_libxkbfile_jll"]
-git-tree-sha1 = "4bcbf660f6c2e714f87e960a171b119d06ee163b"
-uuid = "35661453-b289-5fab-8a00-3d9160c6a3a4"
-version = "1.4.2+4"
-
-[[Xorg_xkeyboard_config_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Xorg_xkbcomp_jll"]
-git-tree-sha1 = "5c8424f8a67c3f2209646d4425f3d415fee5931d"
-uuid = "33bec58e-1273-512f-9401-5d533626f822"
-version = "2.27.0+4"
-
-[[Xorg_xtrans_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "79c31e7844f6ecf779705fbc12146eb190b7d845"
-uuid = "c5fb5394-a638-5e4d-96e5-b29de1b5cf10"
-version = "1.4.0+3"
-
-[[Zlib_jll]]
-deps = ["Libdl"]
-uuid = "83775a58-1f1d-513f-b197-d71354ab007a"
-
-[[Zstd_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "cc4bf3fdde8b7e3e9fa0351bdeedba1cf3b7f6e6"
-uuid = "3161d3a3-bdf6-5164-811a-617609db77b4"
-version = "1.5.0+0"
-
-[[ZygoteRules]]
-deps = ["MacroTools"]
-git-tree-sha1 = "8c1a8e4dfacb1fd631745552c8db35d0deb09ea0"
-uuid = "700de1a5-db45-46bc-99cf-38207098b444"
-version = "0.2.2"
-
-[[libass_jll]]
-deps = ["Artifacts", "Bzip2_jll", "FreeType2_jll", "FriBidi_jll", "HarfBuzz_jll", "JLLWrappers", "Libdl", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "5982a94fcba20f02f42ace44b9894ee2b140fe47"
-uuid = "0ac62f75-1d6f-5e53-bd7c-93b484bb37c0"
-version = "0.15.1+0"
-
-[[libfdk_aac_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "daacc84a041563f965be61859a36e17c4e4fcd55"
-uuid = "f638f0a6-7fb0-5443-88ba-1cc74229b280"
-version = "2.0.2+0"
-
-[[libpng_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Zlib_jll"]
-git-tree-sha1 = "94d180a6d2b5e55e447e2d27a29ed04fe79eb30c"
-uuid = "b53b4c65-9356-5827-b1ea-8c7a1a84506f"
-version = "1.6.38+0"
-
-[[libvorbis_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Ogg_jll", "Pkg"]
-git-tree-sha1 = "c45f4e40e7aafe9d086379e5578947ec8b95a8fb"
-uuid = "f27f6e37-5d2b-51aa-960f-b287f2bc3b7a"
-version = "1.3.7+0"
-
-[[nghttp2_jll]]
-deps = ["Artifacts", "Libdl"]
-uuid = "8e850ede-7688-5339-a07c-302acd2aaf8d"
-
-[[p7zip_jll]]
-deps = ["Artifacts", "Libdl"]
-uuid = "3f19e933-33d8-53b3-aaab-bd5110c3b7a0"
-
-[[x264_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "4fea590b89e6ec504593146bf8b988b2c00922b2"
-uuid = "1270edf5-f2f9-52d2-97e9-ab00b5d0237a"
-version = "2021.5.5+0"
-
-[[x265_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg"]
-git-tree-sha1 = "ee567a171cce03570d77ad3a43e90218e38937a9"
-uuid = "dfaa095f-4041-5dcd-9319-2fabd8486b76"
-version = "3.5.0+0"
-
-[[xkbcommon_jll]]
-deps = ["Artifacts", "JLLWrappers", "Libdl", "Pkg", "Wayland_jll", "Wayland_protocols_jll", "Xorg_libxcb_jll", "Xorg_xkeyboard_config_jll"]
-git-tree-sha1 = "ece2350174195bb31de1a63bea3a41ae1aa593b6"
-uuid = "d8fb68d0-12a3-5cfd-a85a-d49703b185fd"
-version = "0.9.1+5"
-"""
-
-# ╔═╡ Cell order:
-# ╠═9014f3c0-5dec-11ec-1d46-efedaead02f9
-# ╠═e6cfd4c7-fe1f-4991-b5f7-3eaf157b6df9
-# ╠═1bbc8b1b-6ecd-453a-a508-b561ca347ec7
-# ╠═70c668e3-3b70-4f22-a9eb-415203032080
-# ╠═2e687027-42ef-4370-b64a-dd083f37c7a5
-# ╠═e1fafac6-3851-41ba-a4d6-2b78d7d40389
-# ╠═d65aa2b0-166f-467e-b25f-64326be2f22d
-# ╠═4a708fad-4be6-4e80-9aad-338b2bd746ad
-# ╠═9e9b70ee-3244-4688-af1c-0e9d96391c95
-# ╠═c188c227-32da-457a-bb29-2a7a95d31dce
-# ╠═27b0f338-c071-45c6-9b9d-6f910a36748d
-# ╠═0f35870d-b175-4642-b1ce-13904395c7ff
-# ╠═b31ecee6-8bb1-46ae-9360-76fdcf5dfc27
-# ╠═fa5a4623-3bc1-4776-8233-249bdc4bf73f
-# ╠═5bbdc0e8-97ce-4706-a266-4e9df3fbe058
-# ╠═9c67e2ba-1d50-4e63-95bc-dd769bfce93e
-# ╠═cd76be8b-4267-4996-8fc1-0facfa149f97
-# ╠═5c74962a-4b28-4fab-8fc2-e86c6f533cd7
-# ╠═dba4af4a-85ac-457e-ad5c-8d7820a6e942
-# ╟─00000000-0000-0000-0000-000000000001
-# ╟─00000000-0000-0000-0000-000000000002
