+++ src/isorropia/water.jl
@@ -5,7 +5,7 @@ end
 # From Equation 15 in Fountoukis and Nenes (2007), the activity of water is 
 # equal to the relative humidity
 @parameters H2O_aq(t) = 1e-9 [unit = u"mol/m_air^3", isconstantspecies=true,
-                            description = "Dummy water concentration to make units balance (the real concentration is variable `W`)"]
+                            description = "Dummy water concentration to make units balance (the real water concentration is variable `W`)"]
 H2O_aq = ParentScope(H2O_aq)
 H2Oaq = H2O(H2O_aq)
 
@@ -26,32 +26,9 @@ m_aw_coeffs = [
     MgSO4_aqs => [-0.778, 177.74, -719.79, 1174.6, -863.44, 232.31, 0] * unit_molality,
     MgNO32_aqs => [12.166, -16.154, 0, 10.886, 0, -6.815, 0] * unit_molality,
     MgCl2_aqs => [11.505, -26.518, 34.937, -19.829, 0, 0, 0] * unit_molality,
-    # TODO: We are missing some salts here.
+    # TODO(CT): We are missing some salts here.
 ]
 
 # Equation 16.
 W_eq16 = sum([(salt.cation.m / salt.ν_cation) / sum(m_aw_coeff .* RH.^collect(0:6)) 
     for (salt, m_aw_coeff) ∈ m_aw_coeffs])
-
-# Tests
-@test ModelingToolkit.get_unit(W_eq16) == u"kg_water/m_air^3"
-
-@testset "water content" begin
-    ics = Dict([Na_aq => 0, SO4_aq => 10, NH3_aq => 3.4, NO3_aq => 2, Cl_aq => 0, 
-        Ca_aq => 0.4, K_aq => 0.33, Mg_aq => 0.0]) # ug/m3
-    ics = Dict([k => ics[k] / 1e6 / mw[k] for k ∈ keys(ics)]) # ug/m3 / (1e6 ug/g) / g/mol = mol/m3
-    ics[H_aq] = 2*ics[SO4_aq] + ics[NO3_aq] + ics[Cl_aq]
-    w2 = ModelingToolkit.subs_constants(W_eq16)
-    w3 = ModelingToolkit.substitute(w2, ics)
-    RHs = [10, 25, 40, 55, 65, 70, 75, 80, 85, 90] ./ 100.0
-    ws = [Symbolics.value(ModelingToolkit.substitute(w3, [RH => v])) for v ∈ RHs] .*1e9  # kg / m3 * (1e9 ug/kg) = ug/m3
-
-    # TODO(CT): This should be larger than Fountoukis and Nenes (2007) Figure 6a once the rest of the salts are added in.
-    ws_want = [9.274415859681406, 10.232660512301191, 11.34253892156881, 10.55022926762637, 12.790288404478733, 
-    13.626163770467416, 14.659016151077001, 16.101918073323315, 18.42839234123216, 23.167762687858897]
-    @test ws ≈ ws_want
-end
-
-# Test that the activity of water is equal to the relative humidity.
-test_subs = Dict([RH => 0.5])
-@test ModelingToolkit.substitute(activity(H2Oaq), test_subs) == 0.5
